Crate alloy

Alloy

Alloy connects applications to blockchains.

Alloy is a rewrite of ethers-rs from the ground up, with exciting new features, high performance, and excellent docs.

ethers-rs will continue to be maintained until we have achieved feature-parity in Alloy. No action is currently needed from devs.

Overview

This repository contains the following crates:

alloy: Meta-crate for the entire project, including alloy-core

alloy-consensus - Ethereum consensus interface

alloy-contract - Interact with on-chain contracts

alloy-eips - Ethereum Improvement Proprosal (EIP) implementations

alloy-genesis - Ethereum genesis file definitions

alloy-json-rpc - Core data types for JSON-RPC 2.0 clients

alloy-network - Network abstraction for RPC types

alloy-node-bindings - Ethereum execution-layer client bindings

alloy-provider - Interface with an Ethereum blockchain

alloy-pubsub - Ethereum JSON-RPC publish-subscribe tower service and type definitions

alloy-rpc-client - Low-level Ethereum JSON-RPC client implementation

alloy-rpc-types - Ethereum JSON-RPC types

alloy-rpc-engine-types - Ethereum execution-consensus layer (engine) API RPC types

alloy-rpc-trace-types - Ethereum RPC trace types

alloy-signer - Ethereum signer abstraction

alloy-signer-aws - AWS KMS signer implementation

alloy-signer-gcp - GCP KMS signer implementation

alloy-signer-ledger - Ledger signer implementation

alloy-signer-trezor - Trezor signer implementation

alloy-transport - Low-level Ethereum JSON-RPC transport abstraction

alloy-transport-http - HTTP transport implementation

alloy-transport-ipc - IPC transport implementation

alloy-transport-ws - WS transport implementation

Supported Rust Versions

Alloy will keep a rolling MSRV (minimum supported rust version) policy of at least 6 months. When increasing the MSRV, the new Rust version must have been released at least six months ago. The current MSRV is 1.76.

Note that the MSRV is not increased automatically, and only as part of a minor release.

Contributing

Thanks for your help improving the project! We are so happy to have you! We have a contributing guide to help you get involved in the Alloy project.

Pull requests will not be merged unless CI passes, so please ensure that your contribution follows the linting rules and passes clippy.

Note on no_std

Because these crates are primarily network-focused, we do not intend to support no_std for most of them at this time.

The following crates support no_std:

alloy-eips

alloy-genesis

alloy-serde

If you would like to add no_std support to a crate, please make sure to update scripts/check_no_std.sh as well.

Credits

None of these crates would have been possible without the great work done in:

ethers.js

rust-web3

ruint

ethabi

ethcontract-rs

guac_rs

License
Licensed under either of Apache License, Version 2.0 or MIT license at your option.
Unless you explicitly state otherwise, any contribution intentionally submitted for inclusion in these crates by you, as defined in the Apache-2.0 license, shall be dual licensed as above, without any additional terms or conditions.

Re-exports

pub use primitives::hex;

pub use primitives::hex;

pub use primitives::uint;

pub use sol_types::sol;

sol-types

Modules

consensusconsensus

alloy-consensus

contractcontract

alloy-contract

dyn_abidyn-abi

alloy-dyn-abi

eipseips and std

alloy-eips

genesisgenesis

Alloy genesis types

json_abijson-abi

alloy-json-abi

networknetwork

alloy-networks

node_bindingsnode-bindings

alloy-node-bindings

primitives

alloy-primitives

providersproviders

Interface with an Ethereum blockchain.

pubsubpubsub

Ethereum JSON-RPC publish-subscribe tower service and type definitions.

rlprlp

alloy-rlp

rpcrpc

Ethereum JSON-RPC client and types.

signerssigners

Ethereum signer abstraction and implementations.

sol_typessol-types

Solidity type modeling and ABI and EIP-712 codec implementation.

transportstransports

Low-level Ethereum JSON-RPC transport abstraction and implementations.