alloy

Crate consensus

alloy-consensus

Ethereum consensus interface.

This crate contains constants, types, and functions for implementing Ethereum EL consensus and communication. This includes headers, blocks, transactions, EIP-2718 envelopes, EIP-2930, EIP-4844, and more.

In general a type belongs in this crate if it is committed to in the EL block header. This includes:

transactions

blocks

headers

receipts

EIP-2718 envelopes.

Provenance

Much of this code was ported from reth-primitives as part of ongoing alloy migrations.

Modules

conditional

Helpers for conditional transactions.

constants

Ethereum protocol-related constants

crypto

Cryptographic algorithms

error

Helper errors.

extended

Extended transaction types

proofs

Helper function for calculating Merkle proofs and hashes.

transaction

Transaction types.

Structs

Account

Represents an TrieAccount in the account trie.

BlobTransactionSidecar

This represents a set of blobs, and its corresponding commitments and proofs.

Block

Ethereum full block.

BlockBody

A response to GetBlockBodies, containing bodies if any bodies were found.

Header

Ethereum Block header

HeaderInfo

Essential info extracted from a header.

Receipt

Receipt containing result of transaction execution.

ReceiptWithBloom

Receipt with calculated bloom filter.

Receipts

A collection of receipts organized as a two-dimensional vector.

Sealed

A consensus hashable item, with its memoized hash.

SidecarBuilder

Build a BlobTransactionSidecar from an arbitrary amount of data.

Signed

A transaction with a signature and hash seal.

SimpleCoder

Simple coder that only uses the last 31 bytes of each blob. This is the default coder for the SidecarBuilder.

TrieAccount

Represents an TrieAccount in the account trie.

TxEip1559

A transaction with a priority fee (EIP-1559).

TxEip2930

Transaction with an AccessList (EIP-2930).

TxEip4844

EIP-4844 Blob Transaction

TxEip7702

A transaction with a priority fee (EIP-7702).

TxEip4844WithSidecar

EIP-4844 Blob Transaction

TxLegacy

Legacy transaction.

Enums

Eip658Value

Captures the result of a transaction execution.

EthereumTxEnvelope

The Ethereum EIP-2718 Transaction Envelope.

EthereumTypedTransaction

The TypedTransaction enum represents all Ethereum transaction request types.

Extended

An enum that combines two different transaction types.

ReceiptEnvelope

Receipt envelope, as defined in EIP-2718.

TxEip4844Variant

EIP-4844 Blob Transaction

TxType

Transaction types supported by EthereumTxEnvelope.

Constants

EMPTY_OMMER_ROOT_HASH

Ommer root of empty list.

EMPTY_ROOT_HASH

Root hash of an empty trie.

Traits

BlockHeader

Trait for extracting specific Ethereum block data from a header

Eip2718EncodableReceipt

Receipt type that knows its EIP-2718 encoding.

EthBlock

A trait for ethereum like blocks.

RlpDecodableReceipt

Receipt type that knows how to decode itself with a Bloom value.

RlpEncodableReceipt

Receipt type that knows how to encode itself with a Bloom value.

Sealable

Sealeable objects.

SidecarCoder

A strategy for coding and decoding data into sidecars.

SignableTransaction

A signable transaction.

Transaction

Represents a minimal EVM transaction. Currently, EIP-1559, EIP-4844, and EIP-7702 support dynamic fees. We call these transactions “dynamic fee transactions”. We call non dynamic fee transactions(EIP-155, EIP-2930) “legacy fee transactions”.

TransactionEnvelope

A typed transaction envelope.

TxReceipt

Receipt is the result of a transaction execution.

Typed2718

A trait that helps to determine the type of the transaction.

Type Aliases

Blob

A Blob serialized as 0x-prefixed hex string

Bytes48

A commitment/proof serialized as 0x-prefixed hex string

TxEnvelope

The Ethereum EIP-2718 Transaction Envelope.

TypedTransaction

Basic typed transaction which can contain both TxEip4844 and TxEip4844WithSidecar.

Derive Macros

TransactionEnvelope

Derive macro for creating transaction envelope types.