Module alloy::consensus::proofs

Helper function for calculating Merkle proofs and hashes.

Functions

calculate_ommers_root

Calculates the root hash for ommer/uncle headers.

calculate_receipt_root

Calculates the receipt root.

calculate_transaction_root

Calculate a transaction root.

calculate_withdrawals_root

Calculates the root hash of the withdrawals.

ordered_trie_root

Compute a trie root of the collection of rlp encodable items.

ordered_trie_root_with_encoder

Compute a trie root of the collection of items with a custom encoder.

state_root

Calculates the root hash of the state represented as MPT.

state_root_ref_unhashed

Hashes and sorts account keys, then proceeds to calculating the root hash of the state represented as MPT. See state_root_unsorted for more info.

state_root_unhashed

Hashes and sorts account keys, then proceeds to calculating the root hash of the state represented as MPT. See state_root_unsorted for more info.

state_root_unsorted

Sorts the hashed account keys and calculates the root hash of the state represented as MPT. See state_root for more info.

storage_root

Calculates the root hash of account storage trie.

storage_root_unhashed

Hashes storage keys, sorts them and them calculates the root hash of the storage trie. See storage_root_unsorted for more info.

storage_root_unsorted

Sorts and calculates the root hash of account storage trie. See storage_root for more info.