alloy::consensus::proofs

Function state_root_ref_unhashed

pub fn state_root_ref_unhashed<'a, A>(
    state: impl IntoIterator<Item = (&'a Address, &'a A)>,
) -> FixedBytes<32>
where
    A: Into<TrieAccount> + Clone + 'a,

Hashes and sorts account keys, then proceeds to calculating the root hash of the state represented as MPT. See state_root_unsorted for more info.