Trait alloy::consensus::SignableTransaction

pub trait SignableTransaction<Signature>: Transaction {
    // Required methods
    fn set_chain_id(&mut self, chain_id: u64);
    fn encode_for_signing(&self, out: &mut dyn BufMut);
    fn payload_len_for_signature(&self) -> usize;
    fn into_signed(self, signature: Signature) -> Signed<Self, Signature>
       where Self: Sized;

    // Provided methods
    fn set_chain_id_checked(&mut self, chain_id: u64) -> bool { ... }
    fn encoded_for_signing(&self) -> Vec<u8>  { ... }
    fn signature_hash(&self) -> FixedBytes<32> { ... }
}

A signable transaction.

A transaction can have multiple signature types. This is usually alloy_primitives::PrimitiveSignature, however, it may be different for future EIP-2718 transaction types, or in other networks. For example, in Optimism, the deposit transaction signature is the unit type ().

Required Methods

fn set_chain_id(&mut self, chain_id: u64)

Sets chain_id.

Prefer set_chain_id_checked.

fn encode_for_signing(&self, out: &mut dyn BufMut)

RLP-encodes the transaction for signing.

fn payload_len_for_signature(&self) -> usize

Outputs the length of the signature RLP encoding for the transaction.

fn into_signed(self, signature: Signature) -> Signed<Self, Signature>
where Self: Sized,

Convert to a signed transaction by adding a signature and computing the hash.

Provided Methods

fn set_chain_id_checked(&mut self, chain_id: u64) -> bool

Set chain_id if it is not already set. Checks that the provided chain_id matches the existing chain_id if it is already set, returning false if they do not match.

fn encoded_for_signing(&self) -> Vec<u8>

RLP-encodes the transaction for signing it. Used to calculate signature_hash.

See SignableTransaction::encode_for_signing.

fn signature_hash(&self) -> FixedBytes<32>

Calculate the signing hash for the transaction.