alloy::consensus
Module transaction

Transaction types.

Modules

eip4844

[EIP-4844] constants, helpers, and types.

eip4844_utils

Utilities for working with EIP-4844 field elements and implementing SidecarCoder.

Structs

BlobTransactionSidecar

This represents a set of blobs, and its corresponding commitments and proofs.

SidecarBuilder

Build a BlobTransactionSidecar from an arbitrary amount of data.

SimpleCoder

Simple coder that only uses the last 31 bytes of each blob. This is the default coder for the SidecarBuilder.

TxEip1559

A transaction with a priority fee (EIP-1559).

TxEip2930

Transaction with an AccessList (EIP-2930).

TxEip4844

EIP-4844 Blob Transaction

TxEip7702

A transaction with a priority fee (EIP-7702).

TxEip4844WithSidecar

EIP-4844 Blob Transaction

TxLegacy

Legacy transaction.

Enums

TxEip4844Variant

EIP-4844 Blob Transaction

TxEnvelope

The Ethereum EIP-2718 Transaction Envelope.

TxType

Ethereum TransactionType flags as specified in EIPs 2718, 1559, 2930, 4844, and 7702.

TypedTransaction

The TypedTransaction enum represents all Ethereum transaction request types.

Traits

SidecarCoder

A strategy for coding and decoding data into sidecars.

SignableTransaction

A signable transaction.

Transaction

Represents a minimal EVM transaction.

Type Aliases

Blob

A Blob serialized as 0x-prefixed hex string

Bytes48

A commitment/proof serialized as 0x-prefixed hex string