alloy::consensus::transaction

Module signed_legacy_serde

Helper module for encoding signatures of transactions wrapped into Signed in legacy format.

By default, signatures are encoded as a single boolean under yParity key. However, for legacy transactions parity byte is encoded as v key respecting EIP-155 format.

Functions

deserialize

Deserializes signed transaction expecting v key for signature parity.

serialize

Serializes signed transaction with v key for signature parity.