alloy::consensus::transaction

Type Alias PooledTransaction

pub type PooledTransaction = EthereumTxEnvelope<TxEip4844WithSidecar>;

All possible transactions that can be included in a response to GetPooledTransactions. A response to GetPooledTransactions. This can include either a blob transaction, or a non-4844 signed transaction.

The difference between this and the EthereumTxEnvelope<TxEip4844Variant<T>> is that this type always requires the TxEip4844WithSidecar variant, because EIP-4844 transaction can only be propagated with the sidecar over p2p.

Aliased Type

enum PooledTransaction {
    Legacy(Signed<TxLegacy>),
    Eip2930(Signed<TxEip2930>),
    Eip1559(Signed<TxEip1559>),
    Eip4844(Signed<TxEip4844WithSidecar>),
    Eip7702(Signed<TxEip7702>),
}

Variants

Legacy(Signed<TxLegacy>)

An untagged TxLegacy.

Eip2930(Signed<TxEip2930>)

A TxEip2930 tagged with type 1.

Eip1559(Signed<TxEip1559>)

A TxEip1559 tagged with type 2.

Eip4844(Signed<TxEip4844WithSidecar>)

A TxEip4844 tagged with type 3. An EIP-4844 transaction has two network representations: 1 - The transaction itself, which is a regular RLP-encoded transaction and used to retrieve historical transactions..

2 - The transaction with a sidecar, which is the form used to send transactions to the network.

Eip7702(Signed<TxEip7702>)

A TxEip7702 tagged with type 4.