Struct alloy::contract::CallBuilder

pub struct CallBuilder<N, T, P, D>
where
    N: Network,
{
    request: <N as Network>::TransactionRequest,
    block: Option<BlockId>,
    state: Option<HashMap<Address, AccountOverride>>,
    pub provider: P,
    decoder: D,
    transport: PhantomData<T>,
}

Available on crate feature contract only.

A builder for sending a transaction via eth_sendTransaction, or calling a contract via eth_call.

The builder can be .awaited directly, which is equivalent to invoking call. Prefer using call when possible, as awaiting the builder directly will consume it, and currently also boxes the future due to type system limitations.

A call builder can currently be instantiated in the following ways:

by sol!-generated contract structs’ methods (through the #[sol(rpc)] attribute) (SolCallBuilder);

by ContractInstance’s methods (DynCallBuilder);

using CallBuilder::new_raw (RawCallBuilder).

Each method represents a different way to decode the output of the contract call.

Note

This will set state overrides for eth_call, but this is not supported by all clients.

Examples

Using sol!:

use alloy_contract::SolCallBuilder;
use alloy_primitives::{Address, U256};
use alloy_sol_types::sol;

sol! {
    #[sol(rpc)] // <-- Important!
    contract MyContract {
        function doStuff(uint a, bool b) public returns(address c, bytes32 d);
    }
}

let provider = ...;
let address = Address::ZERO;
let contract = MyContract::new(address, &provider);

// Through `contract.<function_name>(args...)`
let a = U256::ZERO;
let b = true;
let builder: SolCallBuilder<_, _, _, MyContract::doStuffCall> = contract.doStuff(a, b);
let MyContract::doStuffReturn { c: _, d: _ } = builder.call().await?;

// Through `contract.call_builder(&<FunctionCall { args... }>)`:
// (note that this is discouraged because it's inherently less type-safe)
let call = MyContract::doStuffCall { a, b };
let builder: SolCallBuilder<_, _, _, MyContract::doStuffCall> = contract.call_builder(&call);
let MyContract::doStuffReturn { c: _, d: _ } = builder.call().await?;

Using ContractInstance:

use alloy_primitives::{Address, Bytes, U256};
use alloy_dyn_abi::DynSolValue;
use alloy_contract::{CallBuilder, ContractInstance, DynCallBuilder, Interface, RawCallBuilder};

let dynamic_abi: JsonAbi = ...;
let interface = Interface::new(dynamic_abi);

let provider = ...;
let address = Address::ZERO;
let contract: ContractInstance<_, _, _> = interface.connect(address, &provider);

// Build and call the function:
let call_builder: DynCallBuilder<_, _, _> = contract.function("doStuff", &[U256::ZERO.into(), true.into()])?;
let result: Vec<DynSolValue> = call_builder.call().await?;

// You can also decode the output manually. Get the raw bytes:
let raw_result: Bytes = call_builder.call_raw().await?;
// Or, equivalently:
let raw_builder: RawCallBuilder<_, _, _> = call_builder.clone().clear_decoder();
let raw_result: Bytes = raw_builder.call().await?;
// Decode the raw bytes:
let decoded_result: Vec<DynSolValue> = call_builder.decode_output(raw_result, false)?;

Fields
request: <N as Network>::TransactionRequestblock: Option<BlockId>state: Option<HashMap<Address, AccountOverride>>provider: P

The provider.

decoder: Dtransport: PhantomData<T>
Implementations

impl<N, T, P> CallBuilder<N, T, P, Function>
where N: Network, T: Transport + Clone, P: Provider<N, T>,

pub fn clear_decoder(self) -> CallBuilder<N, T, P, ()>

Clears the decoder, returning a raw call builder.

impl<N, T, P, C> CallBuilder<N, T, P, PhantomData<C>>
where N: Network, T: Transport + Clone, P: Provider<N, T>, C: SolCall,

pub fn clear_decoder(self) -> CallBuilder<N, T, P, ()>

Clears the decoder, returning a raw call builder.

impl<N, T, P> CallBuilder<N, T, P, ()>
where N: Network, T: Transport + Clone, P: Provider<N, T>,

pub fn new_raw(provider: P, input: Bytes) -> CallBuilder<N, T, P, ()>

Creates a new call builder with the provided provider and ABI encoded input.

Will not decode the output of the call, meaning that call will behave the same as call_raw.

impl<N, T, P, D> CallBuilder<N, T, P, D>
where N: Network, T: Transport + Clone, P: Provider<N, T>, D: CallDecoder,

pub fn from(self, from: Address) -> CallBuilder<N, T, P, D>

Sets the from field in the transaction to the provided value. Defaults to Address::ZERO.

pub fn to(self, to: Option<Address>) -> CallBuilder<N, T, P, D>

Sets the to field in the transaction to the provided address.

pub fn legacy(self) -> CallBuilder<N, T, P, D>

Uses a Legacy transaction instead of an EIP-1559 one to execute the call

pub fn gas(self, gas: Uint<256, 4>) -> CallBuilder<N, T, P, D>

Sets the gas field in the transaction to the provided value

pub fn gas_price(self, gas_price: Uint<256, 4>) -> CallBuilder<N, T, P, D>

Sets the gas_price field in the transaction to the provided value If the internal transaction is an EIP-1559 one, then it sets both max_fee_per_gas and max_priority_fee_per_gas to the same value

pub fn value(self, value: Uint<256, 4>) -> CallBuilder<N, T, P, D>

Sets the value field in the transaction to the provided value

pub fn nonce(self, nonce: u64) -> CallBuilder<N, T, P, D>

Sets the nonce field in the transaction to the provided value

pub fn map<F>(self, f: F) -> CallBuilder<N, T, P, D>
where F: FnOnce(<N as Network>::TransactionRequest) -> <N as Network>::TransactionRequest,

Applies a function to the internal transaction request.

pub const fn block(self, block: BlockId) -> CallBuilder<N, T, P, D>

Sets the block field for sending the tx to the chain

pub fn state( self, state: HashMap<Address, AccountOverride> ) -> CallBuilder<N, T, P, D>

Sets the state override set.

Note

Not all client implementations will support this as a parameter to eth_call.

pub fn calldata(&self) -> &Bytes

Returns the underlying transaction’s ABI-encoded data.

pub async fn estimate_gas(&self) -> Result<Uint<256, 4>, Error>

Returns the estimated gas cost for the underlying transaction to be executed

pub async fn call(&self) -> Result<<D as CallDecoder>::CallOutput, Error>

Queries the blockchain via an eth_call without submitting a transaction to the network.

Returns the decoded the output by using the provided decoder. If this is not desired, use call_raw to get the raw output data.

pub async fn call_raw(&self) -> Result<Bytes, Error>

Queries the blockchain via an eth_call without submitting a transaction to the network.

Does not decode the output of the call, returning the raw output data instead.

See call for more information.

pub fn decode_output( &self, data: Bytes, validate: bool ) -> Result<<D as CallDecoder>::CallOutput, Error>

Decodes the output of a contract function using the provided decoder.

pub async fn deploy(&self) -> Result<Address, Error>

Broadcasts the underlying transaction to the network as a deployment transaction, returning the address of the deployed contract after the transaction has been confirmed.

Returns an error if the transaction is not a deployment transaction, or if the contract address is not found in the deployment transaction’s receipt.

For more fine-grained control over the deployment process, use send instead.

Note that the deployment address can be pre-calculated if the from address and nonce are known using calculate_create_address.

pub async fn send(&self) -> Result<PendingTransactionBuilder<'_, N, T>, Error>

Broadcasts the underlying transaction to the network.

Returns a builder for configuring the pending transaction watcher. See Provider::send_transaction for more information.

pub fn calculate_create_address(&self) -> Option<Address>

Calculates the address that will be created by the transaction, if any.

Returns None if the transaction is not a contract creation (the to field is set), or if the from or nonce fields are not set.

impl<N, T, P, D> CallBuilder<N, T, &P, D>
where N: Network, T: Transport, P: Clone,

pub fn with_cloned_provider(self) -> CallBuilder<N, T, P, D>

Clones the provider and returns a new builder with the cloned provider.