alloy::dyn_abi::abi

Function decode_sequence

pub fn decode_sequence<'de, T>(data: &'de [u8]) -> Result<T, Error>
where
    T: TokenSeq<'de>,

Decodes ABI compliant vector of bytes into vector of tokens described by types param.

You are probably looking for SolValue::abi_decode_sequence if you are not intending to use raw tokens.

See the abi module for more information.