Module alloy::dyn_abi::eip712

Available on crate features dyn-abi and eip712 only.

Implementation of dynamic EIP-712.

This allows for the encoding of EIP-712 messages without having to know the types at compile time. This is useful for things like off-chain signing. It implements the encoding portion of the EIP-712 spec, and does not contain any of the signing logic.

https://eips.ethereum.org/EIPS/eip-712#specification-of-the-eth_signtypeddata-json-rpc

Modules

parser

EIP-712 specific parsing structures.

Structs

Eip712Types

Custom types for TypedData.

PropertyDef

An EIP-712 property definition.

Resolver

A dependency graph built from the Eip712Types object. This is used to safely resolve JSON into a crate::DynSolType by detecting cycles in the type graph and traversing the dep graph.

TypeDef

An EIP-712 type definition.

TypedData

Represents the EIP-712 typed data object.