alloy

Crate eips

alloy-eips

Ethereum Improvement Proprosal (EIP) implementations.

Contains constants, helpers, and basic data structures for consensus EIPs.

EIPs 2930 and 7702 are re-exported from alloy-rs/eips.

Current support

EIP-1559

EIP-1898

EIP-2718

EIP-2930

EIP-2935

EIP-4844

EIP-4788

EIP-4895

EIP-6110

EIP-7002

EIP-7251

EIP-7685

EIP-7702

Modules

eip1559

EIP-1559 constants, helpers, and types.

eip1898

eip2124

Re-export the EIP-2124 types.

eip2718

EIP-2718 traits.

eip2930

Re-export the EIP-2930 types.

eip2935

Contains the history storage contract, first introduced in the Prague hardfork.

eip4788

EIP-4788 constants.

eip4844

EIP-4844 constants and helpers.

eip4895

EIP-4895 Withdrawal type and serde helpers.

eip6110

Contains Deposit request types, first introduced in the Prague hardfork.

eip7002

Contains the system contract and WithdrawalRequest types, first introduced in the Prague hardfork.

eip7251

Contains consolidation types, first introduced in the Prague hardfork.

eip7594

Types and constants for PeerDAS.

eip7623

Constants and utils for calldata cost.

eip7685

EIP-7685: General purpose execution layer requests

eip7691

Contains constants and utility functions for EIP-7691

eip7702

Re-export the EIP-7702 types.

eip7840

Contains constants and utility functions for EIP-7840

eip7892

Contains constants and helper functions for EIP-7892

merge

Constants related to the beacon chain consensus.

Structs

BlobScheduleBlobParams

Blob parameters configuration for a chain, including scheduled updates.

NumHash

A number and a hash.

RpcBlockHash

A block hash which may have a boolean requireCanonical field.

Enums

BlobScheduleEntry

A scheduled blob parameter update entry.

BlockId

A Block Identifier. https://github.com/ethereum/EIPs/blob/master/EIPS/eip-1898.md

BlockNumberOrTag

A block Number (or tag - “latest”, “earliest”, “pending”)

HashOrNumber

Either a hash or a block number

Traits

Decodable2718

Decoding trait for EIP-2718 envelopes. These envelopes wrap a transaction or a receipt with a type flag.

Encodable2718

Encoding trait for EIP-2718 envelopes.

Typed2718

A trait that helps to determine the type of the transaction.

Functions

calc_blob_gasprice

Calculates the blob gas price from the header’s excess blob gas field.

calc_excess_blob_gas

Calculates the excess_blob_gas from the parent header’s blob_gas_used and excess_blob_gas.

calc_next_block_base_fee

Calculate the base fee for the next block based on the EIP-1559 specification.

Type Aliases

BlockHashOrNumber

A block hash or a block number

BlockNumHash

A block number and a hash

ForkBlock

Block number and hash of the forked block.