alloy::eips

Module eip2930

Re-export the EIP-2930 types.

Structs

AccessList

AccessList as defined in EIP-2930

AccessListItem

A list of addresses and storage keys that the transaction plans to access. Accesses outside the list are possible, but become more expensive.

AccessListResult

AccessListResult for handling errors from eth_createAccessList

AccessListWithGasUsed

Access list with gas used appended.