alloy::eips

Module eip4788

EIP-4788 constants.

Constants

BEACON_ROOTS_ADDRESS

The address for the EIP-4788 beacon roots contract.

SYSTEM_ADDRESS

The caller to be used when calling the EIP-4788 beacon roots contract at the beginning of the block.

Statics

BEACON_ROOTS_CODE

The code for the EIP-4788 beacon roots contract.