alloy::eips
Module eip6110

Contains Deposit types, first introduced in the Prague hardfork.

See also EIP-6110: Supply validator deposits on chain

Provides validator deposits as a list of deposit operations added to the Execution Layer block.

Structs

DepositRequest

This structure maps onto the deposit object from EIP-6110.

Constants

MAINNET_DEPOSIT_CONTRACT_ADDRESS

Mainnet deposit contract address.