alloy::eips
Module eip7685

EIP-7685: General purpose execution layer requests

Contains traits for encoding and decoding EIP-7685 requests, as well as validation functions.

Enums

Eip7685Error

EIP-7685 decoding errors.

Traits

Decodable7685

Decoding trait for EIP-7685 requests. The trait should be implemented for an envelope that wraps each possible request type.

Eip7685RequestEnvelope

An EIP-7685 request envelope, blanket implemented for types that impl Encodable7685 and Decodable7685. This envelope is a wrapper around a request, differentiated by the request type.

Encodable7685

Encoding trait for EIP-7685 requests. The trait should be implemented for an envelope that wraps each possible request type.