alloy::eips

Module eip7691

Contains constants and utility functions for EIP-7691

Constants

BLOB_GASPRICE_UPDATE_FRACTION_PECTRA

Determines the maximum rate of change for blob fee after Pectra hardfork activation.

MAX_BLOBS_PER_BLOCK_ELECTRA

CL-enforced maximum blobs per block after Pectra hardfork activation.

TARGET_BLOBS_PER_BLOCK_ELECTRA

CL-enforced target blobs per block after Pectra hardfork activation.

Functions

calc_blob_gasprice

Same as crate::eip4844::calc_blob_gasprice but uses the BLOB_GASPRICE_UPDATE_FRACTION_PECTRA.