Crate alloy::network::primitives

alloy-network-primitives

Primitive types for Alloy network abstraction.

Structs

BlockTransactionHashes

An iterator over the transaction hashes of a block.

Enums

BlockTransactions

Block Transactions depending on the boolean attribute of eth_getBlockBy*, or if used by eth_getUncle*

BlockTransactionsKind

Determines how the transactions field of block should be filled.

Traits

BlockResponse

Block JSON-RPC response.

HeaderResponse

Header JSON-RPC response.

ReceiptResponse

Receipt JSON-RPC response.

TransactionBuilder4844

Transaction builder type supporting EIP-4844 transaction fields.

TransactionBuilder7702

Transaction builder type supporting EIP-7702 transaction fields.

TransactionResponse

Transaction JSON-RPC response. Aggregates transaction data with its block and signer context.