alloy
Crate node_bindings

alloy-node-bindings

Ethereum execution-layer client bindings.

Modules

nodes

Bindings for various nodes.

utils

Utility functions for the node bindings.

Structs

Anvil

Builder for launching anvil.

AnvilInstance

An anvil CLI instance. Will close the instance when dropped.

Geth

Builder for launching geth.

GethInstance

A geth instance. Will close the instance when dropped.

Reth

Builder for launching reth.

RethInstance

A Reth instance. Will close the instance when dropped.

Enums

NodeError

Errors that can occur when working with the node instance.

Constants

EIP1559_FEE_ESTIMATION_DEFAULT_PRIORITY_FEE

The default max priority fee per gas, used in case the base fee is within a threshold.

EIP1559_FEE_ESTIMATION_PAST_BLOCKS

The number of blocks from the past for which the fee rewards are fetched for fee estimation.

EIP1559_FEE_ESTIMATION_PRIORITY_FEE_TRIGGER

The threshold for base fee below which we use the default priority fee, and beyond which we estimate an appropriate value for priority fee.

EIP1559_FEE_ESTIMATION_REWARD_PERCENTILE

The default percentile of gas premiums that are fetched for fee estimation.

EIP1559_FEE_ESTIMATION_THRESHOLD_MAX_CHANGE

The threshold max change/difference (in %) at which we will ignore the fee history values under it.

NODE_DIAL_LOOP_TIMEOUT

Timeout for waiting for the node to add a peer.

NODE_STARTUP_TIMEOUT

How long we will wait for the node to indicate that it is ready.

WEI_IN_ETHER

1 Ether = 1e18 Wei == 0x0de0b6b3a7640000 Wei