Module alloy::node_bindings::nodes::geth

Utilities for launching a Geth dev-mode instance.

Structs

DevOptions

Configuration options that can be set in dev mode.

Geth

Builder for launching geth.

GethInstance

A geth instance. Will close the instance when dropped.

PrivateNetOptions

Configuration options that cannot be set in dev mode.

Enums

NodeMode

Whether or not node is in dev mode and configuration options that depend on the mode.