Module alloy::primitives::postgres

Available on crate feature postgres only.

Support for the [postgres_types] crate.

WARNING: this module depends entirely on [postgres_types, which is not yet stable, therefore this module is exempt from the semver guarantees of this crate.

Enums

FromSqlError

Error when converting from Postgres types.

ToSqlError

Error when converting to Postgres types.