alloy::primitives::ruint::algorithms::div

Function div_nxm

pub fn div_nxm(numerator: &mut [u64], divisor: &mut [u64])

⚠️ In-place Knuth long division with implicit normalization and reciprocals.

⚠️ This function is not part of the stable API.
# Conditions of use:

divisor MUST NOT be empty.

The highest (most-significant) limb of divisor MUST be non-zero.

divisor and numerator MUST contain at least three limbs.

numerator MUST contain at at least as many elements as divisor.

Panics

May panic if any condition of use is violated.