alloy::primitives::ruint::algorithms::div

Function div_nxm_normalized

pub fn div_nxm_normalized(numerator: &mut [u64], divisor: &[u64])

⚠️ In-place Knuth normalized long division with reciprocals.

⚠️ This function is not part of the stable API.
# Conditions of Use

The highest (most-significant) bit of the divisor MUST be set.

The divisor and numerator MUST each be at least two limbs.

numerator MUST contain at at least as many elements as divisor.

Panics

May panic if any condition of use is violated.