Module alloy::providers

Available on crate feature providers only.

Interface with an Ethereum blockchain.

See alloy_provider for more details.

Modules

ext

Extended APIs for the provider module.

fillers

Transaction Fillers

layers

Useful layer implementations for the provider. Currently this module contains the AnvilLayer and AnvilProvider types, when the anvil feature is enabled.

network

alloy-networks

utils

Provider-related utilities.

Structs

Identity

An identity layer that does nothing.

PendingTransaction

Represents a transaction that is yet to be confirmed a specified number of times.

PendingTransactionBuilder

A builder for configuring a pending transaction watcher.

PendingTransactionConfig

Configuration for watching a pending transaction.

ProviderBuilder

A builder for constructing a Provider from various layers.

RootProvider

The root provider manages the RPC client and the heartbeat. It is at the base of every provider stack.

Stack

A stack of two providers.

Traits

Network

Captures type info for network-specific RPC requests/responses.

Provider

Provider is parameterized with a network and a transport. The default transport is type-erased, but you can do Provider<Http, N>.

ProviderLayer

A layering abstraction in the vein of tower::Layer

WalletProvider

Trait for Providers, Fill stacks, etc, which contain NetworkSigner.

Type Aliases

FilterPollerBuilder

A task that polls the provider with eth_getFilterChanges, returning a list of R.

HyperProviderhyper

Type alias for a RootProvider using the Http transport and a hyper client.

ReqwestProviderreqwest

Type alias for a RootProvider using the Http transport and a reqwest client.