alloy::providers::fillers

Trait TxFiller

pub trait TxFiller<N = Ethereum>:
    Clone
    + Send
    + Sync
    + Debug
where
    N: Network,
{
    type Fillable: Send + Sync + 'static;

    // Required methods
    fn status(
        &self,
        tx: &<N as Network>::TransactionRequest,
    ) -> FillerControlFlow;
    fn fill_sync(&self, tx: &mut SendableTx<N>);
    fn prepare<P>(
        &self,
        provider: &P,
        tx: &<N as Network>::TransactionRequest,
    ) -> impl Send + Future<Output = Result<Self::Fillable, RpcError<TransportErrorKind>>>
       where P: Provider<N>;
    fn fill(
        &self,
        fillable: Self::Fillable,
        tx: SendableTx<N>,
    ) -> impl Send + Future<Output = Result<SendableTx<N>, RpcError<TransportErrorKind>>>;

    // Provided methods
    fn join_with<T>(self, other: T) -> JoinFill<Self, T>
       where T: TxFiller<N> { ... }
    fn continue_filling(&self, tx: &SendableTx<N>) -> bool { ... }
    fn ready(&self, tx: &<N as Network>::TransactionRequest) -> bool { ... }
    fn finished(&self, tx: &<N as Network>::TransactionRequest) -> bool { ... }
    fn fill_envelope(
        &self,
        fillable: Self::Fillable,
        tx: SendableTx<N>,
    ) -> impl Send + Future<Output = Result<<N as Network>::TxEnvelope, FillEnvelopeError<<N as Network>::TransactionRequest>>> { ... }
    fn prepare_and_fill<P>(
        &self,
        provider: &P,
        tx: SendableTx<N>,
    ) -> impl Send + Future<Output = Result<SendableTx<N>, RpcError<TransportErrorKind>>>
       where P: Provider<N> { ... }
    fn prepare_call(
        &self,
        tx: &mut <N as Network>::TransactionRequest,
    ) -> impl Send + Future<Output = Result<(), RpcError<TransportErrorKind>>> { ... }
    fn prepare_call_sync(
        &self,
        tx: &mut <N as Network>::TransactionRequest,
    ) -> Result<(), RpcError<TransportErrorKind>> { ... }
}

A layer that can fill in a TransactionRequest with additional information.

Lifecycle Notes

The FillerControlFlow determines the lifecycle of a filler. Fillers may be in one of three states:

Missing: The filler is missing a required property to fill in the transaction request. TxFiller::status should return FillerControlFlow::Missing. with a list of the missing properties.

Ready: The filler is ready to fill in the transaction request. TxFiller::status should return FillerControlFlow::Ready.

Finished: The filler has filled in all properties that it can fill. TxFiller::status should return FillerControlFlow::Finished.

Required Associated Types

type Fillable: Send + Sync + 'static

The properties that this filler retrieves from the RPC. to fill in the TransactionRequest.

Required Methods

fn status(&self, tx: &<N as Network>::TransactionRequest) -> FillerControlFlow

Return a control-flow enum indicating whether the filler is ready to fill in the transaction request, or if it is missing required properties.

fn fill_sync(&self, tx: &mut SendableTx<N>)

Performs any synchronous filling. This should be called before TxFiller::prepare and TxFiller::fill to fill in any properties that can be filled synchronously.

fn prepare<P>( &self, provider: &P, tx: &<N as Network>::TransactionRequest, ) -> impl Send + Future<Output = Result<Self::Fillable, RpcError<TransportErrorKind>>>
where P: Provider<N>,

Prepares fillable properties, potentially by making an RPC request.

fn fill( &self, fillable: Self::Fillable, tx: SendableTx<N>, ) -> impl Send + Future<Output = Result<SendableTx<N>, RpcError<TransportErrorKind>>>

Fills in the transaction request with the fillable properties.

Provided Methods

fn join_with<T>(self, other: T) -> JoinFill<Self, T>
where T: TxFiller<N>,

Joins this filler with another filler to compose multiple fillers.

fn continue_filling(&self, tx: &SendableTx<N>) -> bool

Returns true if the filler should continue filling.

fn ready(&self, tx: &<N as Network>::TransactionRequest) -> bool

Returns true if the filler is ready to fill in the transaction request.

fn finished(&self, tx: &<N as Network>::TransactionRequest) -> bool

Returns true if the filler is finished filling in the transaction request.

fn fill_envelope( &self, fillable: Self::Fillable, tx: SendableTx<N>, ) -> impl Send + Future<Output = Result<<N as Network>::TxEnvelope, FillEnvelopeError<<N as Network>::TransactionRequest>>>

Fills in the transaction request and try to convert it to an envelope.

fn prepare_and_fill<P>( &self, provider: &P, tx: SendableTx<N>, ) -> impl Send + Future<Output = Result<SendableTx<N>, RpcError<TransportErrorKind>>>
where P: Provider<N>,

Prepares and fills the transaction request with the fillable properties.

fn prepare_call( &self, tx: &mut <N as Network>::TransactionRequest, ) -> impl Send + Future<Output = Result<(), RpcError<TransportErrorKind>>>

Prepares transaction request with necessary fillers required for eth_call operations asynchronously

fn prepare_call_sync( &self, tx: &mut <N as Network>::TransactionRequest, ) -> Result<(), RpcError<TransportErrorKind>>

Prepares transaction request with necessary fillers required for eth_call operations synchronously

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.