alloy::providers

Module layers

Useful layer implementations for the provider.

Structs

AnvilLayer

A layer that wraps an Anvil config.

AnvilProvider

A provider that wraps an AnvilInstance, preventing the instance from being dropped while the provider is in use.

CacheLayer

A provider layer that caches RPC responses and serves them on subsequent requests.

CacheProvider

The CacheProvider holds the underlying in-memory LRU cache and overrides methods from the Provider trait. It attempts to fetch from the cache and fallbacks to the RPC in case of a cache miss.

CallBatchLayer

Provider layer that aggregates contract calls (eth_call) over a time period into a single Multicall3 contract call.

CallBatchProvider

A provider that batches multiple requests into a single request.

ChainLayer

A layer that wraps a NamedChain. The layer will be used to set the client’s poll interval based on the average block time for this chain.

SharedCache

Shareable cache.