Struct alloy::providers::layers::AnvilProvider

pub struct AnvilProvider<P, T> {
    inner: P,
    _anvil: Arc<AnvilInstance>,
    _pd: PhantomData<fn() -> T>,
}

Available on crate feature providers only.

A provider that wraps an AnvilInstance, preventing the instance from being dropped while the provider is in use.

Fields
inner: P_anvil: Arc<AnvilInstance>_pd: PhantomData<fn() -> T>
Implementations

impl<P, T> AnvilProvider<P, T>
where P: Provider<T>, T: Transport + Clone,

pub fn new(inner: P, _anvil: Arc<AnvilInstance>) -> AnvilProvider<P, T>

Creates a new AnvilProvider with the given inner provider and anvil instance.