alloy::providers::network::any

Struct AnyRpcBlock

pub struct AnyRpcBlock(pub WithOtherFields<Block<AnyRpcTransaction, Header<AnyHeader>>>);

A wrapper for AnyRpcBlock that allows for handling unknown block types.

This type wraps:

rpc transaction

additional fields

Tuple Fields
0: WithOtherFields<Block<AnyRpcTransaction, Header<AnyHeader>>>
Implementations

impl AnyRpcBlock

pub const fn new( inner: WithOtherFields<Block<AnyRpcTransaction, Header<AnyHeader>>>, ) -> AnyRpcBlock

Create a new AnyRpcBlock.

pub fn into_inner(self) -> Block<AnyRpcTransaction, Header<AnyHeader>>

Consumes the type and returns the wrapped rpc block.

pub fn try_into_consensus<T, H>(self) -> Result<Block<T, H>, AnyConversionError>
where T: TryFrom<AnyRpcTransaction>, <T as TryFrom<AnyRpcTransaction>>::Error: Error + Send + Sync + 'static, H: TryFrom<AnyHeader>, <H as TryFrom<AnyHeader>>::Error: Error + Send + Sync + 'static,

Attempts to convert the inner RPC Block into a consensus block.

Returns an AnyConversionError if any of the conversions fail.

pub fn try_into_transactions( self, ) -> Result<Vec<AnyRpcTransaction>, ValueError<BlockTransactions<AnyRpcTransaction>>>

Tries to convert inner transactions into a vector of AnyRpcTransaction.

Returns an error if the block contains only transaction hashes or if it is an uncle block.

pub fn into_transactions_iter(self) -> impl Iterator<Item = AnyRpcTransaction>

Consumes the type and returns an iterator over the transactions in this block