alloy::providers::network

Enum AnyTxEnvelope

pub enum AnyTxEnvelope {
    Ethereum(EthereumTxEnvelope<TxEip4844Variant>),
    Unknown(UnknownTxEnvelope),
}

Transaction envelope for a catch-all network.

Variants

Ethereum(EthereumTxEnvelope<TxEip4844Variant>)

An Ethereum transaction.

Unknown(UnknownTxEnvelope)

A transaction with unknown type.

Implementations

impl AnyTxEnvelope

pub const fn is_ethereum(&self) -> bool

Returns true if this is the ethereum transaction variant

pub const fn is_unknown(&self) -> bool

Returns true if this is the unknown transaction variant

pub const fn as_envelope(&self) -> Option<&EthereumTxEnvelope<TxEip4844Variant>>

Returns the inner Ethereum transaction envelope, if it is an Ethereum transaction.

pub const fn as_unknown(&self) -> Option<&UnknownTxEnvelope>

Returns the inner UnknownTxEnvelope if it is an unknown variant.

pub fn try_into_envelope( self, ) -> Result<EthereumTxEnvelope<TxEip4844Variant>, ValueError<AnyTxEnvelope>>

Returns the inner Ethereum transaction envelope, if it is an Ethereum transaction. If the transaction is not an Ethereum transaction, it is returned as an error.

pub fn try_into_unknown(self) -> Result<UnknownTxEnvelope, AnyTxEnvelope>

Returns the inner UnknownTxEnvelope, if it is an unknown transaction. If the transaction is not an unknown transaction, it is returned as an error.

pub fn try_into_either<T>( self, ) -> Result<Either<EthereumTxEnvelope<TxEip4844Variant>, T>, <T as TryFrom<UnknownTxEnvelope>>::Error>
where T: TryFrom<UnknownTxEnvelope>,

Attempts to convert the UnknownTxEnvelope into Either::Right if this is an unknown variant.

Returns Either::Left with the ethereum TxEnvelope if this is the AnyTxEnvelope::Ethereum variant and Either::Right with the converted variant.

pub fn try_map_unknown<T, E>( self, f: impl FnOnce(UnknownTxEnvelope) -> Result<T, E>, ) -> Result<Either<EthereumTxEnvelope<TxEip4844Variant>, T>, E>

Attempts to convert the UnknownTxEnvelope into Either::Right if this is an AnyTxEnvelope::Unknown.

Returns Either::Left with the ethereum TxEnvelope if this is the AnyTxEnvelope::Ethereum variant and Either::Right with the converted variant.

pub const fn as_legacy(&self) -> Option<&Signed<TxLegacy>>

Returns the TxLegacy variant if the transaction is a legacy transaction.

pub const fn as_eip2930(&self) -> Option<&Signed<TxEip2930>>

Returns the TxEip2930 variant if the transaction is an EIP-2930 transaction.

pub const fn as_eip1559(&self) -> Option<&Signed<TxEip1559>>

Returns the TxEip1559 variant if the transaction is an EIP-1559 transaction.

pub const fn as_eip4844(&self) -> Option<&Signed<TxEip4844Variant>>

Returns the TxEip4844Variant variant if the transaction is an EIP-4844 transaction.

pub const fn as_eip7702(&self) -> Option<&Signed<TxEip7702>>

Returns the TxEip7702 variant if the transaction is an EIP-7702 transaction.