Trait alloy::providers::network::primitives::ReceiptResponse

pub trait ReceiptResponse {

    // Required methods
    fn contract_address(&self) -> Option<Address>;
    fn status(&self) -> bool;
    fn block_hash(&self) -> Option<FixedBytes<32>>;
    fn block_number(&self) -> Option<u64>;
    fn transaction_hash(&self) -> FixedBytes<32>;
    fn transaction_index(&self) -> Option<u64>;
    fn gas_used(&self) -> u64;
    fn effective_gas_price(&self) -> u128;
    fn blob_gas_used(&self) -> Option<u64>;
    fn blob_gas_price(&self) -> Option<u128>;
    fn from(&self) -> Address;
    fn to(&self) -> Option<Address>;
    fn authorization_list(&self) -> Option<&[SignedAuthorization]>;
    fn cumulative_gas_used(&self) -> u64;
    fn state_root(&self) -> Option<FixedBytes<32>>;

}

Receipt JSON-RPC response.

Required Methods

fn contract_address(&self) -> Option<Address>

Address of the created contract, or None if the transaction was not a deployment.

fn status(&self) -> bool

Status of the transaction.

Note

Caution must be taken when using this method for deep-historical receipts, as it may not accurately reflect the status of the transaction. The transaction status is not knowable from the receipt for transactions before [EIP-658].

fn block_hash(&self) -> Option<FixedBytes<32>>

Hash of the block this transaction was included within.

fn block_number(&self) -> Option<u64>

Number of the block this transaction was included within.

fn transaction_hash(&self) -> FixedBytes<32>

Transaction Hash.

fn transaction_index(&self) -> Option<u64>

Index within the block.

fn gas_used(&self) -> u64

Gas used by this transaction alone.

fn effective_gas_price(&self) -> u128

Effective gas price.

fn blob_gas_used(&self) -> Option<u64>

Blob gas used by the eip-4844 transaction.

fn blob_gas_price(&self) -> Option<u128>

Blob gas price paid by the eip-4844 transaction.

fn from(&self) -> Address

Address of the sender.

fn to(&self) -> Option<Address>

Address of the receiver.

fn authorization_list(&self) -> Option<&[SignedAuthorization]>

EIP-7702 Authorization list.

fn cumulative_gas_used(&self) -> u64

Returns the cumulative gas used at this receipt.

fn state_root(&self) -> Option<FixedBytes<32>>

The post-transaction state root (pre Byzantium)

EIP98 makes this field optional.