Trait alloy::providers::network::TransactionBuilder

pub trait TransactionBuilder<N>: Default + Sized + Send + Sync + 'static
where
    N: Network,
{

    // Required methods
    fn chain_id(&self) -> Option<u64>;
    fn set_chain_id(&mut self, chain_id: u64);
    fn nonce(&self) -> Option<u64>;
    fn set_nonce(&mut self, nonce: u64);
    fn input(&self) -> Option<&Bytes>;
    fn set_input(&mut self, input: Bytes);
    fn from(&self) -> Option<Address>;
    fn set_from(&mut self, from: Address);
    fn to(&self) -> Option<TxKind>;
    fn set_to(&mut self, to: TxKind);
    fn value(&self) -> Option<Uint<256, 4>>;
    fn set_value(&mut self, value: Uint<256, 4>);
    fn gas_price(&self) -> Option<Uint<256, 4>>;
    fn set_gas_price(&mut self, gas_price: Uint<256, 4>);
    fn max_fee_per_gas(&self) -> Option<Uint<256, 4>>;
    fn set_max_fee_per_gas(&mut self, max_fee_per_gas: Uint<256, 4>);
    fn max_priority_fee_per_gas(&self) -> Option<Uint<256, 4>>;
    fn set_max_priority_fee_per_gas(
        &mut self,
        max_priority_fee_per_gas: Uint<256, 4>
    );
    fn max_fee_per_blob_gas(&self) -> Option<Uint<256, 4>>;
    fn set_max_fee_per_blob_gas(&mut self, max_fee_per_blob_gas: Uint<256, 4>);
    fn gas_limit(&self) -> Option<Uint<256, 4>>;
    fn set_gas_limit(&mut self, gas_limit: Uint<256, 4>);
    fn get_blob_sidecar(&self) -> Option<&BlobTransactionSidecar>;
    fn set_blob_sidecar(&mut self, sidecar: BlobTransactionSidecar);
    fn build_unsigned(
        self
    ) -> Result<<N as Network>::UnsignedTx, TransactionBuilderError>;
    fn build<S>(
        self,
        signer: &S
    ) -> impl Send + Future<Output = Result<<N as Network>::TxEnvelope, TransactionBuilderError>>
       where S: NetworkSigner<N>;

    // Provided methods
    fn with_chain_id(self, chain_id: u64) -> Self { ... }
    fn with_nonce(self, nonce: u64) -> Self { ... }
    fn with_input(self, input: Bytes) -> Self { ... }
    fn with_from(self, from: Address) -> Self { ... }
    fn with_to(self, to: TxKind) -> Self { ... }
    fn calculate_create_address(&self) -> Option<Address> { ... }
    fn with_value(self, value: Uint<256, 4>) -> Self { ... }
    fn with_gas_price(self, gas_price: Uint<256, 4>) -> Self { ... }
    fn with_max_fee_per_gas(self, max_fee_per_gas: Uint<256, 4>) -> Self { ... }
    fn with_max_priority_fee_per_gas(
        self,
        max_priority_fee_per_gas: Uint<256, 4>
    ) -> Self { ... }
    fn with_max_fee_per_blob_gas(
        self,
        max_fee_per_blob_gas: Uint<256, 4>
    ) -> Self { ... }
    fn with_gas_limit(self, gas_limit: Uint<256, 4>) -> Self { ... }
    fn with_blob_sidecar(self, sidecar: BlobTransactionSidecar) -> Self { ... }

}

Available on crate feature providers only.

A Transaction builder for a network.

Transaction builders are primarily used to construct typed transactions that can be signed with TransactionBuilder::build, or unsigned typed transactions with TransactionBuilder::build_unsigned.

Transaction builders should be able to construct all available transaction types on a given network.

Required Methods

fn chain_id(&self) -> Option<u64>

Get the chain ID for the transaction.

fn set_chain_id(&mut self, chain_id: u64)

Set the chain ID for the transaction.

fn nonce(&self) -> Option<u64>

Get the nonce for the transaction.

fn set_nonce(&mut self, nonce: u64)

Set the nonce for the transaction.

fn input(&self) -> Option<&Bytes>

Get the input data for the transaction.

fn set_input(&mut self, input: Bytes)

Set the input data for the transaction.

fn from(&self) -> Option<Address>

Get the sender for the transaction.

fn set_from(&mut self, from: Address)

Set the sender for the transaction.

fn to(&self) -> Option<TxKind>

Get the recipient for the transaction.

fn set_to(&mut self, to: TxKind)

Set the recipient for the transaction.

fn value(&self) -> Option<Uint<256, 4>>

Get the value for the transaction.

fn set_value(&mut self, value: Uint<256, 4>)

Set the value for the transaction.

fn gas_price(&self) -> Option<Uint<256, 4>>

Get the legacy gas price for the transaction.

fn set_gas_price(&mut self, gas_price: Uint<256, 4>)

Set the legacy gas price for the transaction.

fn max_fee_per_gas(&self) -> Option<Uint<256, 4>>

Get the max fee per gas for the transaction.

fn set_max_fee_per_gas(&mut self, max_fee_per_gas: Uint<256, 4>)

Set the max fee per gas for the transaction.

fn max_priority_fee_per_gas(&self) -> Option<Uint<256, 4>>

Get the max priority fee per gas for the transaction.

fn set_max_priority_fee_per_gas( &mut self, max_priority_fee_per_gas: Uint<256, 4> )

Set the max priority fee per gas for the transaction.

fn max_fee_per_blob_gas(&self) -> Option<Uint<256, 4>>

Get the max fee per blob gas for the transaction.

fn set_max_fee_per_blob_gas(&mut self, max_fee_per_blob_gas: Uint<256, 4>)

Set the max fee per blob gas for the transaction.

fn gas_limit(&self) -> Option<Uint<256, 4>>

Get the gas limit for the transaction.

fn set_gas_limit(&mut self, gas_limit: Uint<256, 4>)

Set the gas limit for the transaction.

fn get_blob_sidecar(&self) -> Option<&BlobTransactionSidecar>

Gets the EIP-4844 blob sidecar of the transaction.

fn set_blob_sidecar(&mut self, sidecar: BlobTransactionSidecar)

Sets the EIP-4844 blob sidecar of the transaction.

Note: This will also set the versioned blob hashes accordingly: BlobTransactionSidecar::versioned_hashes

fn build_unsigned( self ) -> Result<<N as Network>::UnsignedTx, TransactionBuilderError>

Build an unsigned, but typed, transaction.

fn build<S>( self, signer: &S ) -> impl Send + Future<Output = Result<<N as Network>::TxEnvelope, TransactionBuilderError>>
where S: NetworkSigner<N>,

Build a signed transaction.

Provided Methods

fn with_chain_id(self, chain_id: u64) -> Self

Builder-pattern method for setting the chain ID.

fn with_nonce(self, nonce: u64) -> Self

Builder-pattern method for setting the nonce.

fn with_input(self, input: Bytes) -> Self

Builder-pattern method for setting the input data.

fn with_from(self, from: Address) -> Self

Builder-pattern method for setting the sender.

fn with_to(self, to: TxKind) -> Self

Builder-pattern method for setting the recipient.

fn calculate_create_address(&self) -> Option<Address>

Calculates the address that will be created by the transaction, if any.

Returns None if the transaction is not a contract creation (the to field is set), or if the from or nonce fields are not set.

fn with_value(self, value: Uint<256, 4>) -> Self

Builder-pattern method for setting the value.

fn with_gas_price(self, gas_price: Uint<256, 4>) -> Self

Builder-pattern method for setting the legacy gas price.

fn with_max_fee_per_gas(self, max_fee_per_gas: Uint<256, 4>) -> Self

Builder-pattern method for setting max fee per gas .

fn with_max_priority_fee_per_gas( self, max_priority_fee_per_gas: Uint<256, 4> ) -> Self

Builder-pattern method for setting max priority fee per gas.

fn with_max_fee_per_blob_gas(self, max_fee_per_blob_gas: Uint<256, 4>) -> Self

Builder-pattern method for setting max fee per blob gas .

fn with_gas_limit(self, gas_limit: Uint<256, 4>) -> Self

Builder-pattern method for setting the gas limit.

fn with_blob_sidecar(self, sidecar: BlobTransactionSidecar) -> Self

Builder-pattern method for setting the EIP-4844 blob sidecar of the transaction.

Object Safety

This trait is not object safe.