alloy::providers

Trait MulticallItem

pub trait MulticallItem {
    type Decoder: SolCall;

    // Required methods
    fn value(&self) -> Uint<256, 4>;
    fn target(&self) -> Address;
    fn input(&self) -> Bytes;

    // Provided method
    fn into_call(self, allow_failure: bool) -> CallItem<Self::Decoder>
       where Self: Sized { ... }
}

A trait that is to be implemented by a type that can be distilled to a singular contract call item.

Required Associated Types

type Decoder: SolCall

Decoder for the return data of the call.

Required Methods

fn value(&self) -> Uint<256, 4>

Returns the value to send with the call.

fn target(&self) -> Address

The target address of the call.

fn input(&self) -> Bytes

ABI-encoded input data for the call.

Provided Methods

fn into_call(self, allow_failure: bool) -> CallItem<Self::Decoder>
where Self: Sized,

Converts self to a CallItem while specifying whether it can fail.