alloy::providers

Module utils

Provider-related utilities.

Structs

Eip1559Estimation

Return type of EIP1155 gas fee estimator.

Enums

Eip1559Estimator

EIP-1559 estimator variants

Constants

EIP1559_BASE_FEE_MULTIPLIER

Multiplier for the current base fee to estimate max base fee for the next block.

EIP1559_FEE_ESTIMATION_PAST_BLOCKS

The number of blocks from the past for which the fee rewards are fetched for fee estimation.

EIP1559_FEE_ESTIMATION_REWARD_PERCENTILE

The default percentile of gas premiums that are fetched for fee estimation.

EIP1559_MIN_PRIORITY_FEE

The minimum priority fee to provide.

Traits

Eip1559EstimatorFn

A trait responsible for estimating EIP-1559 values

Functions

eip1559_default_estimator

The default EIP-1559 fee estimator.

Type Aliases

EstimatorFunction

An estimator function for EIP1559 fees.

JoinedRecommendedFillers

Helper type representing the joined recommended fillers i.e GasFiller, BlobGasFiller, NonceFiller, and ChainIdFiller.