Crate alloy::rlp

Available on crate feature rlp only.

alloy-rlp

This crate provides Ethereum RLP (de)serialization functionality. RLP is commonly used for Ethereum EL datastructures, and its documentation can be found at ethereum.org.

Usage

We strongly recommend deriving RLP traits via the RlpEncodable and RlpDecodable derive macros.

Trait methods can then be accessed via the Encodable and Decodable traits.

Example

use alloy_rlp::{RlpEncodable, RlpDecodable, Decodable, Encodable};

#[derive(Debug, RlpEncodable, RlpDecodable, PartialEq)]
pub struct MyStruct {
    pub a: u64,
    pub b: Vec<u8>,
}

let my_struct = MyStruct {
    a: 42,
    b: vec![1, 2, 3],
};

let mut buffer = Vec::<u8>::new();
let encoded = my_struct.encode(&mut buffer);
let decoded = MyStruct::decode(&mut buffer.as_slice()).unwrap();
assert_eq!(my_struct, decoded);

Modules

bytes

Provides abstractions for working with bytes.

Macros

impl_max_encoded_len

Implement MaxEncodedLen and MaxEncodedLenAssoc for a type.

Structs

Bytes

A cheaply cloneable and sliceable chunk of contiguous memory.

BytesMut

A unique reference to a contiguous slice of memory.

Header

The header of an RLP item.

Rlp

An active RLP decoder, with a specific slice of a payload.

Enums

Error

RLP error type.

Constants

EMPTY_LIST_CODE

RLP prefix byte for a 0-length array.

EMPTY_STRING_CODE

RLP prefix byte for 0-length string.

Traits

Buf

Read bytes from a buffer.

BufMut

A trait for values that provide sequential write access to bytes.

Decodable

A type that can be decoded from an RLP blob.

Encodable

A type that can be encoded via RLP.

MaxEncodedLen

Defines the max length of an Encodable type as a const generic.

MaxEncodedLenAssoc

Defines the max length of an Encodable type as an associated constant.

Functions

encode

Encode a value.

encode_fixed_sizearrayvec

Encode a type with a known maximum size.

encode_iter

Encode all items from an iterator.

encode_list

Encode a list of items.

length_of_length

Determine the length in bytes of the length prefix of an RLP item.

list_length

Calculate the length of a list.

Type Aliases

Result

RLP result type.

Derive Macros

RlpDecodable

Derives Decodable for the type whose implementation expects an rlp-list input: <rlp-header, fields...>

RlpDecodableWrapper

Derives Decodable for the type whose implementation expects only the individual fields encoded: <fields...>

RlpEncodable

Derives Encodable for the type which encodes the all fields as list: <rlp-header, fields...>

RlpEncodableWrapper

Derives Encodable for the type which encodes the fields as-is, without a header: <fields...>

RlpMaxEncodedLen

Derives MaxEncodedLen for types of constant size.