Trait alloy::rlp::MaxEncodedLenAssoc

pub unsafe trait MaxEncodedLenAssoc: Encodable {
    const LEN: usize;
}

Available on crate feature rlp only.

Defines the max length of an Encodable type as an associated constant.

Safety

An invalid value can cause the encoder to panic.

Required Associated Constants

const LEN: usize

The maximum length.

Object Safety

This trait is not object safe.