alloy::rpc

Crate client

alloy-rpc-client

Low-level Ethereum JSON-RPC client implementation.

Usage

Usage of this crate typically means instantiating an RpcClient over some Transport. The RPC client can then be used to make requests to the RPC server. Requests are captured as RpcCall futures, which can then be polled to completion.

For example, to make a simple request:

// Instantiate a new client over a transport.
let client: ReqwestClient = ClientBuilder::default().http(url);

// Prepare a request to the server.
let request = client.request_noparams("eth_blockNumber");

// Poll the request to completion.
let block_number: U64 = request.await.unwrap();

Batch requests are also supported:

// Instantiate a new client over a transport.
let client: ReqwestClient = ClientBuilder::default().http(url);

// Prepare a batch request to the server.
let mut batch = client.new_batch();

// Batches serialize params immediately. So we need to handle the result when
// adding calls.
let block_number_fut = batch.add_call("eth_blockNumber", &()).unwrap();
let balance_fut = batch.add_call("eth_getBalance", address).unwrap();

// Make sure to send the batch!
batch.send().await.unwrap();

// After the batch is complete, we can get the results.
// Note that requests may error separately!
let block_number: U64 = block_number_fut.await.unwrap();
let balance: U256 = balance_fut.await.unwrap();

Structs

BatchRequest

A batch JSON-RPC request, used to bundle requests into a single transport call.

ClientBuilder

A builder for the transport RpcClient.

PollChannel

A channel yielding responses from a poller task.

PollerBuilder

A poller task builder.

PollerStream

A stream of responses from polling an RPC method.

RpcCall

A prepared, but unsent, RPC call.

RpcClient

A JSON-RPC client.

RpcClientInner

A JSON-RPC client.

Waiter

Awaits a single response for a request that has been included in a batch.

Enums

BuiltInConnectionString

Connection string for built-in transports.

Type Aliases

ClientRef

A borrowed RpcClient.

NoParams

Parameter type of a JSON-RPC request with no parameters.

ReqwestClient

A client using a reqwest HTTP transport.

WeakClient

An RpcClient in a Weak reference.