Struct alloy::rpc::client::PollerBuilder

pub struct PollerBuilder<Conn, Params, Resp> {
    client: Weak<RpcClientInner<Conn>>,
    method: Cow<'static, str>,
    params: Params,
    channel_size: usize,
    poll_interval: Duration,
    limit: usize,
    _pd: PhantomData<fn() -> Resp>,
}

Available on crate features rpc and rpc-client only.

A poller task builder.

This builder is used to create a poller task that repeatedly polls a method on a client and sends the responses to a channel. By default, this is done every 10 seconds, with a channel size of 16, and no limit on the number of successful polls. This is all configurable.

The builder is consumed using the spawn method, which returns a channel to receive the responses. The task will continue to poll until either the client or the channel is dropped.

The channel can be converted into a stream using the into_stream method.

Alternatively, into_stream can be used to directly return a stream of responses on the current thread. This is currently equivalent to spawn().into_stream(), but this may change in the future.

Examples

Poll eth_blockNumber every 5 seconds:

use alloy_primitives::U64;
use alloy_rpc_client::PollerBuilder;
use futures_util::StreamExt;

let poller: PollerBuilder<_, (), U64> = client
    .prepare_static_poller("eth_blockNumber", ())
    .with_poll_interval(std::time::Duration::from_secs(5));
let mut stream = poller.into_stream();
while let Some(block_number) = stream.next().await {
   println!("polled block number: {block_number}");
}

Fields
client: Weak<RpcClientInner<Conn>>method: Cow<'static, str>params: Paramschannel_size: usizepoll_interval: Durationlimit: usize_pd: PhantomData<fn() -> Resp>
Implementations

impl<Conn, Params, Resp> PollerBuilder<Conn, Params, Resp>
where Conn: Transport + Clone, Params: RpcParam + 'static, Resp: RpcReturn + Clone,

pub fn new( client: Weak<RpcClientInner<Conn>>, method: impl Into<Cow<'static, str>>, params: Params, ) -> PollerBuilder<Conn, Params, Resp>

Available on crate feature providers only.

Create a new poller task.

pub const fn channel_size(&self) -> usize

Available on crate feature providers only.

Returns the channel size for the poller task.

pub fn set_channel_size(&mut self, channel_size: usize)

Available on crate feature providers only.

Sets the channel size for the poller task.

pub fn with_channel_size( self, channel_size: usize, ) -> PollerBuilder<Conn, Params, Resp>

Available on crate feature providers only.

Sets the channel size for the poller task.

pub const fn limit(&self) -> usize

Available on crate feature providers only.

Returns the limit on the number of successful polls.

pub fn set_limit(&mut self, limit: Option<usize>)

Available on crate feature providers only.

Sets a limit on the number of successful polls.

pub fn with_limit( self, limit: Option<usize>, ) -> PollerBuilder<Conn, Params, Resp>

Available on crate feature providers only.

Sets a limit on the number of successful polls.

pub const fn poll_interval(&self) -> Duration

Available on crate feature providers only.

Returns the duration between polls.

pub fn set_poll_interval(&mut self, poll_interval: Duration)

Available on crate feature providers only.

Sets the duration between polls.

pub fn with_poll_interval( self, poll_interval: Duration, ) -> PollerBuilder<Conn, Params, Resp>

Available on crate feature providers only.

Sets the duration between polls.

pub fn spawn(self) -> PollChannel<Resp>

Available on crate feature providers only.

Starts the poller in a new Tokio task, returning a channel to receive the responses on.

pub fn into_stream(self) -> impl Stream<Item = Resp> + Unpin

Available on crate feature providers only.

Starts the poller and returns the stream of responses.

Note that this is currently equivalent to self.spawn().into_stream(), but this may change in the future.