Struct alloy::rpc::client::RpcClientInner

pub struct RpcClientInner<T> { /* private fields */ }

A JSON-RPC client.

This struct manages a Transport and a request ID counter. It is used to build RpcCall and BatchRequest objects. The client delegates transport access to the calls.

Note

IDs are allocated sequentially, starting at 0. IDs are reserved via RpcClientInner::next_id. Note that allocated IDs may not be used. There is no guarantee that a prepared RpcCall will be sent, or that a sent call will receive a response.

Implementations

impl<T> RpcClientInner<T>

pub const fn new(t: T, is_local: bool) -> RpcClientInner<T>

Create a new RpcClient with the given transport.

Note: Sets the poll interval to 250ms for local transports and 7s for remote transports by default.

pub fn with_id(self, id: u64) -> RpcClientInner<T>

Sets the starting ID for the client.

pub fn poll_interval(&self) -> Duration

Returns the default poll interval (milliseconds) for the client.

pub fn set_poll_interval(&self, poll_interval: Duration)

Set the poll interval for the client in milliseconds. Default: 7s for remote and 250ms for local transports.

pub const fn transport(&self) -> &T

Returns a reference to the underlying transport.

pub fn transport_mut(&mut self) -> &mut T

Returns a mutable reference to the underlying transport.

pub fn into_transport(self) -> T

Consumes the client and returns the underlying transport.

pub fn make_request<Params>( &self, method: impl Into<Cow<'static, str>>, params: Params, ) -> Request<Params>
where Params: RpcParam,

Build a JsonRpcRequest with the given method and params.

This function reserves an ID for the request, however the request is not sent.

To send a request, use RpcClientInner::request and await the returned RpcCall.

pub const fn is_local(&self) -> bool

true if the client believes the transport is local.

This can be used to optimize remote API usage, or to change program behavior on local endpoints. When the client is instantiated by parsing a URL or other external input, this value is set on a best-efforts basis and may be incorrect.

pub fn set_local(&mut self, is_local: bool)

Set the is_local flag.

pub fn next_id(&self) -> Id

Reserve a request ID u64.

impl<T> RpcClientInner<T>
where T: Transport + Clone,

pub fn request<Params, Resp>( &self, method: impl Into<Cow<'static, str>>, params: Params, ) -> RpcCall<T, Params, Resp>
where Params: RpcParam, Resp: RpcReturn,

Prepares an RpcCall.

This function reserves an ID for the request, however the request is not sent. To send a request, await the returned RpcCall.

Note

Serialization is done lazily. It will not be performed until the call is awaited. This means that if a serializer error occurs, it will not be caught until the call is awaited.

pub fn request_noparams<Resp>( &self, method: impl Into<Cow<'static, str>>, ) -> RpcCall<T, [(); 0], Resp>
where Resp: RpcReturn,

Prepares an RpcCall with no parameters.

See request for more details.

pub fn boxed(self) -> RpcClientInner<BoxTransport>

Type erase the service in the transport, allowing it to be used in a generic context.

Note:

This is for abstracting over RpcClient<T> for multiple T by erasing each type. E.g. if you have RpcClient<Http> and RpcClient<Ws> you can put both into a Vec<RpcClient<BoxTransport>>.