Crate alloy::rpc::types

Available on crate features rpc and rpc-types only.

Ethereum JSON-RPC type definitions.

alloy-rpc-types

Meta-crate for all Ethereum JSON-RPC types.

Modules

admin

alloy-rpc-types-admin

anvil

alloy-rpc-types-anvil

beacon

alloy-rpc-types-beacon

engine

alloy-rpc-types-engine

eth

alloy-rpc-types-eth

serde_helpers

Alloy serde helpers for primitive types.

trace

alloy-rpc-types-trace

txpool

alloy-rpc-types-txpool

Structs

AccessList

AccessList as defined in EIP-2930

AccessListItem

A list of addresses and storage keys that the transaction plans to access. Accesses outside the list are possible, but become more expensive.

AccessListWithGasUsed

Access list with gas used appended.

AccountInfo

Account information.

AnyReceiptEnvelope

Receipt envelope, as defined in EIP-2718.

BlobTransactionSidecar

This represents a set of blobs, and its corresponding commitments and proofs.

Block

Block representation

BlockNumHash

Block number and hash.

BlockOverrides

BlockOverrides is a set of header fields to override.

BlockTransactionHashes

An iterator over the transaction hashes of a block.

BlockTransactionHashesMut

An Iterator over the transaction hashes of a block.

BloomFilter

Helper type to represent a bloom filter used for matching logs.

Bundle

Bundle of transactions

ChainStatus

Chain status.

EIP1186AccountProofResponse

Response for EIP-1186 account proof eth_getProof

EIP1186StorageProof

Data structure with proof for one single storage-entry

EthCallResponse

CallResponse for eth_callMany

ExtAccountInfo

Extended account information (used by parity_allAccountInfo).

FeeHistory

Response type for eth_feeHistory

Filter

Filter for logs.

FilterSet

FilterSet is a set of values that will be used to filter logs

FilteredParams

Support for matching Filters

Header

Block header representation.

Index

A hex encoded or decimal index that’s intended to be used as a rust index, hence it’s deserialized into a usize.

Log

Ethereum Log emitted by a transaction

OptimismTransactionReceiptFields

Additional fields for Optimism transaction receipts

Parity

Type that represents the signature parity byte, meant for use in RPC.

PeerEthProtocolInfo

Peer Ethereum protocol information

PeerInfo

Peer connection information

PeerNetworkInfo

Peer network information

PeerProtocolsInfo

Peer protocols information

Peers

Peers info

PipProtocolInfo

Peer PIP protocol information

RawLog

Ethereum Log

Receipt

Receipt containing result of transaction execution.

ReceiptWithBloom

Receipt with calculated bloom filter.

RecoveredAccount

account derived from a signature as well as information that tells if it is valid for the current chain

Rich

Value representation with additional info

RpcBlockHash

A block hash which may have a boolean requireCanonical field. If false, an RPC call should raise if a block matching the hash is not found. If true, an RPC call should additionally raise if the block is not in the canonical chain. https://github.com/ethereum/EIPs/blob/master/EIPS/eip-1898.md#specification

Signature

Container type for all signature fields in RPC

StateContext

State context for callMany

SyncInfo

Syncing info

Transaction

Transaction object used in RPC

TransactionInfo

Additional fields in the context of a block that contains this transaction.

TransactionInput

Helper type that supports both data and input fields that map to transaction input data.

TransactionReceipt

Transaction receipt

TransactionRequest

Represents all transaction requests to/from RPC.

TransactionStats

Propagation statistics for pending transaction.

TxGasAndReward

Internal struct to calculate reward percentiles

Withdrawal

Withdrawal represents a validator withdrawal from the consensus layer.

Work

The result of an eth_getWork request

Enums

BlockError

Error that can occur when converting other types to blocks

BlockHashOrNumber

Either a block hash or a block number

BlockId

A Block Identifier. https://github.com/ethereum/EIPs/blob/master/EIPS/eip-1898.md

BlockNumberOrTag

A block Number (or tag - “latest”, “earliest”, “pending”)

BlockTransactions

Block Transactions depending on the boolean attribute of eth_getBlockBy*, or if used by eth_getUncle*

BlockTransactionsKind

Determines how the transactions field of Block should be filled.

ConversionError

Error variants when converting from crate::Transaction to alloy_consensus::Signed transaction.

FilterBlockOption

Represents the target range of blocks for the filter

FilterChanges

Response of the eth_getFilterChanges RPC.

FilterId

Owned equivalent of a SubscriptionId

PeerCount

Number of peers connected to.

PendingTransactionFilterKind

Specifies the kind of information you wish to receive from the eth_newPendingTransactionFilter RPC endpoint.

ReceiptEnvelope

Receipt envelope, as defined in EIP-2718.

SyncStatus

Sync status

ValueOrArray

Union type for representing a single value or a vector of values inside a filter

Functions

calc_blob_gasprice

Calculates the blob gas price from the header’s excess blob gas field.

calc_excess_blob_gas

Calculates the excess_blob_gas from the parent header’s blob_gas_used and excess_blob_gas.

logs_bloom

Calculate receipt logs bloom.

Type Aliases

AnyTransactionReceipt

Alias for a catch-all receipt type.

ForkBlock

Block number and hash of the forked block.

RichBlock

A Block representation that allows to include additional fields

RichHeader

Header representation with additional info.

Topic

A single topic