alloy::rpc::types::eth::erc4337

Type Alias ConditionalOptions

pub type ConditionalOptions = TransactionConditional;

๐Deprecated since 0.8.4: use TransactionConditional instead

Alias for backwards compat

Aliased Type

struct ConditionalOptions {
    pub known_accounts: HashMap<Address, AccountStorage, FbBuildHasher<20>>,
    pub block_number_min: Option<u64>,
    pub block_number_max: Option<u64>,
    pub timestamp_min: Option<u64>,
    pub timestamp_max: Option<u64>,
}

Fields
known_accounts: HashMap<Address, AccountStorage, FbBuildHasher<20>>

A map of account addresses to their expected storage states. Each account can have a specified storage root or explicit slot-value pairs.

block_number_min: Option<u64>

The minimal block number at which the transaction can be included. None indicates no minimum block number constraint.

block_number_max: Option<u64>

The maximal block number at which the transaction can be included. None indicates no maximum block number constraint.

timestamp_min: Option<u64>

The minimal timestamp at which the transaction can be included. None indicates no minimum timestamp constraint.

timestamp_max: Option<u64>

The maximal timestamp at which the transaction can be included. None indicates no maximum timestamp constraint.