Struct alloy::rpc::types::Transaction

pub struct Transaction<T = TxEnvelope> {
    pub inner: T,
    pub block_hash: Option<FixedBytes<32>>,
    pub block_number: Option<u64>,
    pub transaction_index: Option<u64>,
    pub effective_gas_price: Option<u128>,
    pub from: Address,
}

Transaction object used in RPC

Fields
inner: T

The inner transaction object

block_hash: Option<FixedBytes<32>>

Hash of block where transaction was included, None if pending

block_number: Option<u64>

Number of block where transaction was included, None if pending

transaction_index: Option<u64>

Transaction Index

effective_gas_price: Option<u128>

Deprecated effective gas price value.

from: Address

Sender

Implementations

impl<T> Transaction<T>
where T: Transaction,

pub fn is_legacy_gas(&self) -> bool

Returns true if the transaction is a legacy or 2930 transaction.

impl<T> Transaction<T>
where T: Transaction + Encodable2718,

pub fn info(&self) -> TransactionInfo

Returns the TransactionInfo for this transaction.

This contains various metadata about the transaction and block context if available.

impl<T> Transaction<T>
where T: Into<TransactionRequest>,

pub fn into_request(self) -> TransactionRequest

Converts Transaction into TransactionRequest.

During this conversion data for TransactionRequest::sidecar is not populated as it is not part of Transaction.