Module alloy::rpc::types::trace::geth

Available on crate features rpc and rpc-types and rpc-types-trace only.

Geth tracing types.

Modules

call

Geth call tracer types.

four_byte

Geth 4byte tracer types.

mux

Geth muxTracer types.

noop

Noop tracer response.

pre_state

Pre-state Geth tracer types.

Structs

AccountState

Represents the state of an account

BlockTraceResult

blockTraceResult represents the results of tracing a single block when an entire chain is being traced. ref https://github.com/ethereum/go-ethereum/blob/ee530c0d5aa70d2c00ab5691a89ab431b73f8165/eth/tracers/api.go#L218-L222

CallConfig

The configuration for the call tracer.

CallFrame

The response object for debug_traceTransaction with "tracer": "callTracer".

CallLogFrame

Represents a recorded call.

DefaultFrame

Geth Default struct log trace frame

DiffMode

Represents the account states before and after the transaction is executed.

FourByteFrame

The 4byte tracer response object.

GethDebugTracerConfig

Configuration of the tracer

GethDebugTracingCallOptions

Bindings for additional debug_traceCall options

GethDebugTracingOptions

Bindings for additional debug_traceTransaction options

GethDefaultTracingOptions

Default tracing options for the struct looger.

NoopFrame

An empty frame response that’s only an empty json object {}.

PreStateConfig

The config for the prestate tracer.

PreStateMode

Includes all the account states necessary to execute a given transaction.

StructLog

Represents a struct log entry in a trace

Enums

AccountChangeKind

Helper type to track the kind of change of an AccountState.

DiffStateKind

Helper type for DiffMode to represent a specific set

GethDebugBuiltInTracerType

Available built-in tracers

GethDebugTracerType

Available tracers

GethTrace

Tracing response objects

PreStateFrame

A tracer that records AccountStates. The prestate tracer has two modes: prestate and diff

Type Aliases

TraceResult

Result type for geth style transaction trace