alloy

Module signers

Ethereum signer abstraction and implementations.

See alloy_signer for more details.

Modules

k256

RustCrypto: secp256k1 (K-256) elliptic curve

local

alloy-signer-local

utils

Utility functions for working with Ethereum signatures.

Macros

sign_transaction_with_chain_id

Utility to get and set the chain ID on a transaction and the resulting signature within a signer’s sign_transaction.

Structs

Signature

An Ethereum ECDSA signature.

Enums

Either

The enum Either with variants Left and Right is a general purpose sum type with two cases.

Error

Generic error type for Signer implementations.

UnsupportedSignerOperation

An unsupported signer operation.

Traits

Signer

Asynchronous Ethereum signer.

SignerSync

Synchronous Ethereum signer.

Type Aliases

Result

Result type alias for Error.