alloy::signers::k256::ecdsa::hazmat

Trait VerifyPrimitive

pub trait VerifyPrimitive<C>:
    Sized
    + AffineCoordinates<FieldRepr = GenericArray<u8, <C as Curve>::FieldBytesSize>>
    + Copy
where
    C: PrimeCurve + CurveArithmetic<AffinePoint = Self>,
    <<C as Curve>::FieldBytesSize as Add>::Output: ArrayLength<u8>,
{
    // Provided methods
    fn verify_prehashed(
        &self,
        z: &GenericArray<u8, <C as Curve>::FieldBytesSize>,
        sig: &Signature<C>,
    ) -> Result<(), Error> { ... }
    fn verify_digest<D>(
        &self,
        msg_digest: D,
        sig: &Signature<C>,
    ) -> Result<(), Error>
       where D: FixedOutput<OutputSize = <C as Curve>::FieldBytesSize> { ... }
}

Verify the given prehashed message using ECDSA.

This trait is intended to be implemented on type which can access the affine point represeting the public key via &self, such as a particular curve’s AffinePoint type.

Provided Methods

fn verify_prehashed( &self, z: &GenericArray<u8, <C as Curve>::FieldBytesSize>, sig: &Signature<C>, ) -> Result<(), Error>

Verify the prehashed message against the provided ECDSA signature.

Accepts the following arguments:

z: message digest to be verified. MUST BE OUTPUT OF A CRYPTOGRAPHICALLY SECURE DIGEST ALGORITHM!!!

sig: signature to be verified against the key and message

fn verify_digest<D>( &self, msg_digest: D, sig: &Signature<C>, ) -> Result<(), Error>
where D: FixedOutput<OutputSize = <C as Curve>::FieldBytesSize>,

Verify message digest against the provided signature.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.