alloy::signers::k256::ecdsa::signature::digest::block_buffer::generic_array
Module sequence

Useful traits for manipulating sequences of data stored in GenericArrays

Traits

Concat

Defines GenericSequences which can be joined together, forming a larger array.

GenericSequence

Defines some sequence with an associated length and iteration capabilities.

Lengthen

Defines any GenericSequence which can be lengthened or extended by appending or prepending an element to it.

Shorten

Defines a GenericSequence which can be shortened by removing the first or last element from it.

Split

Defines a GenericSequence that can be split into two parts at a given pivot index.

Type Aliases

SequenceItem

Accessor for GenericSequence item type, which is really IntoIterator::Item