alloy::signers::k256::ecdsa::signature::digest::block_buffer::generic_array::typenum::marker_traits

Trait TypeArray

pub trait TypeArray: Sealed { }

The marker trait for type-level arrays of type-level numbers.

Someday, it may contain an associated constant to produce a runtime array, like the other marker traits here. However, that is blocked by this issue.