alloy::signers::k256::ecdsa::signature::digest::block_buffer::generic_array::typenum

Trait SquareRoot

pub trait SquareRoot {
    type Output;
}

A type operator for taking the integer square root of Self.

The integer square root of n is the largest integer m such that n >= m*m. This definition is equivalent to truncating the real-valued square root: floor(real_sqrt(n)).

Required Associated Types

type Output

The result of the integer square root.