alloy::signers::k256::ecdsa::signature::digest::const_oid

Type Alias Arc

pub type Arc = u32;

Type alias used to represent an “arc” (i.e. integer identifier value).

X.660 does not define a maximum size of an arc.

The current representation is u32, which has been selected as being sufficient to cover the current PKCS/PKIX use cases this library has been used in conjunction with.

Future versions may potentially make it larger if a sufficiently important use case is discovered.