alloy::signers::k256::ecdsa::signature::digest::core_api

Trait FixedOutputCore

pub trait FixedOutputCore:
    UpdateCore
    + BufferKindUser
    + OutputSizeUser
where
    Self::BlockSize: IsLess<UInt<UInt<UInt<UInt<UInt<UInt<UInt<UInt<UInt<UTerm, B1>, B0>, B0>, B0>, B0>, B0>, B0>, B0>, B0>>,
    <Self::BlockSize as IsLess<UInt<UInt<UInt<UInt<UInt<UInt<UInt<UInt<UInt<UTerm, B1>, B0>, B0>, B0>, B0>, B0>, B0>, B0>, B0>>>::Output: NonZero,
{
    // Required method
    fn finalize_fixed_core(
        &mut self,
        buffer: &mut BlockBuffer<Self::BlockSize, Self::BufferKind>,
        out: &mut GenericArray<u8, Self::OutputSize>,
    );
}

Core trait for hash functions with fixed output size.

Required Methods

fn finalize_fixed_core( &mut self, buffer: &mut BlockBuffer<Self::BlockSize, Self::BufferKind>, out: &mut GenericArray<u8, Self::OutputSize>, )

Finalize state using remaining data stored in the provided block buffer, write result into provided array and leave self in a dirty state.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.