alloy::signers::k256::elliptic_curve::bigint

Module prelude

Import prelude for this crate: includes important traits.

Traits

AddMod

Compute self + rhs mod p.

ArrayDecoding

Support for decoding a GenericArray as a big integer.

ArrayEncoding

Support for encoding a big integer as a GenericArray.

Bounded

Integers whose representation takes a bounded amount of space.

CheckedAdd

Checked addition.

CheckedMul

Checked multiplication.

CheckedSub

Checked subtraction.

Concat

Concatenate two numbers into a “wide” double-width value, using the lo value as the least significant value.

ConcatMixed

Concatenate two numbers into a “wide” combined-width value, using the lo value as the least significant value.

Encoding

Encoding support.

Integer

Integer type.

Invert

Constant-time inversion.

MulMod

Compute self * rhs mod p.

MultiExponentiate

Performs modular multi-exponentiation using Montgomery’s ladder.

MultiExponentiateBoundedExp

Performs modular multi-exponentiation using Montgomery’s ladder. exponent_bits represents the number of bits to take into account for the exponent.

NegMod

Compute -self mod p.

Pow

Constant-time exponentiation.

PowBoundedExp

Constant-time exponentiation with exponent of a bounded bit size.

Random

Random number generation support.

RandomMod

Modular random number generation support.

Split

Split a number in half, returning the most significant half followed by the least significant.

SplitMixed

Split a number into parts, returning the most significant part followed by the least significant.

Square

Support for optimized squaring

SubMod

Compute self - rhs mod p.

Zero

Zero values.