alloy::signers::k256::elliptic_curve::bigint::prelude

Trait Split

pub trait Split: SplitMixed<Self::Output, Self::Output> {
    type Output;

    // Provided method
    fn split(&self) -> (Self::Output, Self::Output) { ... }
}

Split a number in half, returning the most significant half followed by the least significant.

Required Associated Types

type Output

Split output: high/low components of the value.

Provided Methods

fn split(&self) -> (Self::Output, Self::Output)

Split this number in half, returning its high and low components respectively.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.