Trait alloy::signers::k256::elliptic_curve::ff::BatchInvert

pub trait BatchInvert<F>
where
    F: Field,
{
    // Required method
    fn batch_invert(self) -> F;
}

Available on crate features signers and alloc only.

Extension trait for iterators over mutable field elements which allows those field elements to be inverted in a batch.

I: IntoIterator<Item = &'a mut F: Field + ConstantTimeEq> implements this trait when the alloc feature flag is enabled.

For non-allocating contexts, see the BatchInverter struct.

Required Methods

fn batch_invert(self) -> F

Consumes this iterator and inverts each field element (when nonzero). Zero-valued elements are left as zero.

Returns the inverse of the product of all nonzero field elements.