alloy::signers::k256::elliptic_curve::group

Module cofactor

Traits

CofactorCurve

Efficient representation of an elliptic curve point guaranteed to be in the correct prime order subgroup.

CofactorCurveAffine

Affine representation of an elliptic curve point guaranteed to be in the correct prime order subgroup.

CofactorGroup

This trait represents an element of a cryptographic group with a large prime-order subgroup and a comparatively-small cofactor.