alloy::signers::k256::elliptic_curve::group
Module prime

Traits

PrimeCurve

Efficient representation of an elliptic curve point guaranteed to be in the correct prime order subgroup.

PrimeCurveAffine

Affine representation of an elliptic curve point guaranteed to be in the correct prime order subgroup.

PrimeGroup

This trait represents an element of a prime-order cryptographic group.