Trait alloy::signers::k256::elliptic_curve::pkcs8::der::Decode

pub trait Decode<'a>: Sized {
    // Required method
    fn decode<R>(decoder: &mut R) -> Result<Self, Error>
       where R: Reader<'a>;

    // Provided method
    fn from_der(bytes: &'a [u8]) -> Result<Self, Error> { ... }
}

Decoding trait.

This trait provides the core abstraction upon which all decoding operations are based.

Required Methods

fn decode<R>(decoder: &mut R) -> Result<Self, Error>
where R: Reader<'a>,

Attempt to decode this message using the provided decoder.

Provided Methods

fn from_der(bytes: &'a [u8]) -> Result<Self, Error>

Parse Self from the provided DER-encoded byte slice.

Object Safety

This trait is not object safe.