Struct alloy::signers::k256::elliptic_curve::point::NonIdentity

pub struct NonIdentity<P> {
    point: P,
}

Available on crate feature signers only.

Non-identity point type.

This type ensures that its value is not the identity point, ala core::num::NonZero*.

In the context of ECC, it’s useful for ensuring that certain arithmetic cannot result in the identity point.

Fields
point: P
Implementations

impl<P> NonIdentity<P>
where P: ConditionallySelectable + ConstantTimeEq + Default,

pub fn new(point: P) -> CtOption<NonIdentity<P>>

Create a NonIdentity from a point.

impl<P> NonIdentity<P>
where P: ConditionallySelectable + ConstantTimeEq + Default + GroupEncoding,

pub fn from_repr(repr: &<P as GroupEncoding>::Repr) -> CtOption<NonIdentity<P>>

Decode a NonIdentity from its encoding.

impl<P> NonIdentity<P>
where P: Copy,

pub fn to_point(self) -> P

Return wrapped point.

impl<P> NonIdentity<P>
where P: ConditionallySelectable + ConstantTimeEq + Curve + Default,

pub fn random(rng: impl CryptoRng + RngCore) -> NonIdentity<P>

Generate a random NonIdentity<ProjectivePoint>.

pub fn to_affine(self) -> NonIdentity<<P as Curve>::AffineRepr>

Converts this element into its affine representation.

impl<P> NonIdentity<P>
where P: PrimeCurveAffine,

pub fn to_curve(self) -> NonIdentity<<P as PrimeCurveAffine>::Curve>

Converts this element to its curve representation.