alloy::signers::k256::elliptic_curve::point

Trait DecompressPoint

pub trait DecompressPoint<C>: Sized
where
    C: Curve,
{
    // Required method
    fn decompress(
        x: &GenericArray<u8, <C as Curve>::FieldBytesSize>,
        y_is_odd: Choice,
    ) -> CtOption<Self>;
}

Decompress an elliptic curve point.

Point decompression recovers an original curve point from its x-coordinate and a boolean flag indicating whether or not the y-coordinate is odd.

Required Methods

fn decompress( x: &GenericArray<u8, <C as Curve>::FieldBytesSize>, y_is_odd: Choice, ) -> CtOption<Self>

Attempt to decompress an elliptic curve point.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.