Struct alloy::signers::k256::elliptic_curve::SecretKey

pub struct SecretKey<C>
where
    C: Curve,
{
    inner: ScalarPrimitive<C>,
}

Available on crate feature signers only.

Elliptic curve secret keys.

This type wraps a secret scalar value, helping to prevent accidental exposure and securely erasing the value from memory when dropped.

Parsing PKCS#8 Keys

PKCS#8 is a commonly used format for encoding secret keys (especially ones generated by OpenSSL).

Keys in PKCS#8 format are either binary (ASN.1 BER/DER), or PEM encoded (ASCII) and begin with the following:

-----BEGIN PRIVATE KEY-----

To decode an elliptic curve private key from PKCS#8, enable the pkcs8 feature of this crate (or the pkcs8 feature of a specific RustCrypto elliptic curve crate) and use the [DecodePrivateKey] trait to parse it.

When the pem feature of this crate (or a specific RustCrypto elliptic curve crate) is enabled, a [FromStr] impl is also available.

Fields
inner: ScalarPrimitive<C>
Implementations

impl<C> SecretKey<C>
where C: Curve,

pub fn random(rng: &mut impl CryptoRngCore) -> SecretKey<C>
where C: CurveArithmetic,

Available on crate feature arithmetic only.

Generate a random SecretKey.

pub fn new(scalar: ScalarPrimitive<C>) -> SecretKey<C>

Create a new secret key from a scalar value.

pub fn as_scalar_primitive(&self) -> &ScalarPrimitive<C>

Borrow the inner secret ScalarPrimitive value.

⚠️ Warning

This value is key material.

Please treat it with the care it deserves!

pub fn to_nonzero_scalar(&self) -> NonZeroScalar<C>
where C: CurveArithmetic,

Available on crate feature arithmetic only.

Get the secret NonZeroScalar value for this key.

⚠️ Warning

This value is key material.

Please treat it with the care it deserves!

pub fn public_key(&self) -> PublicKey<C>
where C: CurveArithmetic,

Available on crate feature arithmetic only.

Get the PublicKey which corresponds to this secret key

pub fn from_bytes( bytes: &GenericArray<u8, <C as Curve>::FieldBytesSize>, ) -> Result<SecretKey<C>, Error>

Deserialize secret key from an encoded secret scalar.

pub fn from_slice(slice: &[u8]) -> Result<SecretKey<C>, Error>

Deserialize secret key from an encoded secret scalar passed as a byte slice.

The slice is expected to be a minimum of 24-bytes (192-byts) and at most C::FieldBytesSize bytes in length.

Byte slices shorter than the field size are handled by zero padding the input.

pub fn to_bytes(&self) -> GenericArray<u8, <C as Curve>::FieldBytesSize>

Serialize raw secret scalar as a big endian integer.

pub fn from_sec1_der(der_bytes: &[u8]) -> Result<SecretKey<C>, Error>
where C: Curve + ValidatePublicKey, <C as Curve>::FieldBytesSize: ModulusSize,

Available on crate feature sec1 only.

Deserialize secret key encoded in the SEC1 ASN.1 DER ECPrivateKey format.

pub fn to_sec1_der(&self) -> Result<Zeroizing<Vec<u8>>, Error>
where C: CurveArithmetic, <C as CurveArithmetic>::AffinePoint: FromEncodedPoint<C> + ToEncodedPoint<C>, <C as Curve>::FieldBytesSize: ModulusSize,

Available on crate features alloc and arithmetic and sec1 only.

Serialize secret key in the SEC1 ASN.1 DER ECPrivateKey format.

pub fn from_sec1_pem(s: &str) -> Result<SecretKey<C>, Error>
where C: Curve + ValidatePublicKey, <C as Curve>::FieldBytesSize: ModulusSize,

Available on crate feature pem only.

Parse SecretKey from PEM-encoded SEC1 ECPrivateKey format.

PEM-encoded SEC1 keys can be identified by the leading delimiter:

-----BEGIN EC PRIVATE KEY-----

pub fn to_sec1_pem( &self, line_ending: LineEnding, ) -> Result<Zeroizing<String>, Error>
where C: CurveArithmetic, <C as CurveArithmetic>::AffinePoint: FromEncodedPoint<C> + ToEncodedPoint<C>, <C as Curve>::FieldBytesSize: ModulusSize,

Available on crate feature pem only.

Serialize private key as self-zeroizing PEM-encoded SEC1 ECPrivateKey with the given pem::LineEnding.

Pass Default::default() to use the OS’s native line endings.