Struct alloy::signers::k256::Scalar

pub struct Scalar(pub(crate) Uint<crypto_bigint::::uint::U256::{constant#0}>);

Available on crate feature signers only.

Scalars are elements in the finite field modulo n.

Trait impls

Much of the important functionality of scalars is provided by traits from the ff crate, which is re-exported as k256::elliptic_curve::ff:

Field - represents elements of finite fields and provides:

Field::random - generate a random scalar

double, square, and invert operations

Bounds for Add, Sub, Mul, and Neg (as well as *Assign equivalents)

Bounds for ConditionallySelectable from the subtle crate

PrimeField - represents elements of prime fields and provides:

from_repr/to_repr for converting field elements from/to big integers.

multiplicative_generator and root_of_unity constants.

PrimeFieldBits - operations over field elements represented as bits (requires bits feature)

Please see the documentation for the relevant traits for more information.

serde support

When the serde feature of this crate is enabled, the Serialize and Deserialize traits are impl’d for this type.

The serialization is a fixed-width big endian encoding. When used with textual formats, the binary data is encoded as hexadecimal.

Tuple Fields
0: Uint<crypto_bigint::::uint::U256::{constant#0}>
Implementations

impl Scalar

pub const ZERO: Scalar = _

Zero scalar.

pub const ONE: Scalar = _

Multiplicative identity.

pub fn is_zero(&self) -> Choice

Checks if the scalar is zero.

pub fn to_bytes(&self) -> GenericArray<u8, <Secp256k1 as Curve>::FieldBytesSize>

Returns the SEC1 encoding of this scalar.

pub const fn negate(&self) -> Scalar

Negates the scalar.

pub const fn add(&self, rhs: &Scalar) -> Scalar

Returns self + rhs mod n.

pub const fn sub(&self, rhs: &Scalar) -> Scalar

Returns self - rhs mod n.

pub fn mul(&self, rhs: &Scalar) -> Scalar

Modulo multiplies two scalars.

pub fn square(&self) -> Scalar

Modulo squares the scalar.

pub fn shr_vartime(&self, shift: usize) -> Scalar

Right shifts the scalar.

Note: not constant-time with respect to the shift parameter.

pub fn invert(&self) -> CtOption<Scalar>

Inverts the scalar.

pub fn generate_biased(rng: &mut impl CryptoRngCore) -> Scalar

Returns a (nearly) uniformly-random scalar, generated in constant time.

pub fn generate_vartime(rng: &mut impl RngCore) -> Scalar

Returns a uniformly-random scalar, generated using rejection sampling.