Crate alloy::signers::local

Available on crate features signers and signer-local only.

alloy-signer-local

Local signer implementations:

K256 private key

Mnemonic phrase

YubiHSM2

Features

keystore: enables Ethereum keystore functionality on the PrivateKeySigner type.

mnemonic: enables BIP-39 mnemonic functionality for building PrivateKeySigners.

yubihsm: enables LocalSigners with YubiHSM2 support.

Modules

coins_bip39

The bip39 crate is heavily inspired by and reuses code from Wagyu under the MIT license. The difference being, the underlying extended private keys are generated using the bip32 crate, that depends on k256 instead of libsecp256k1.

yubihsm

Structs

LocalSigner

An Ethereum private-public key pair which can be used for signing messages.

MnemonicBuilder

Represents a structure that can resolve into a PrivateKeySigner.

Enums

LocalSignerError

Error thrown by LocalSigner.

Type Aliases

PrivateKeySigner

A signer instantiated with a locally stored private key.

YubiSigneryubihsm

A signer instantiated with a YubiHSM.