Trait alloy::signers::local::coins_bip39::Wordlist

pub trait Wordlist {
    // Required method
    fn get_all() -> &'static [&'static str];

    // Provided methods
    fn get(index: usize) -> Result<&'static str, WordlistError> { ... }
    fn get_index(word: &str) -> Result<usize, WordlistError> { ... }
}

Available on crate features signers and signer-local only.

The Wordlist trait that every language’s wordlist must implement.

Required Methods

fn get_all() -> &'static [&'static str]

Returns the word list as a string.

Implementor’s note: this MUST be sorted

Provided Methods

fn get(index: usize) -> Result<&'static str, WordlistError>

Returns the word of a given index from the word list.

fn get_index(word: &str) -> Result<usize, WordlistError>

Returns the index of a given word from the word list.

Object Safety

This trait is not object safe.