Module alloy::signers::local::yubihsm::client

Available on crate features signers and signer-local only.

YubiHSM client: core functionality of this crate.

The Client type provides a set of methods which map to commands which interface with the HSM.

https://developers.yubico.com/YubiHSM2/Commands/

Structs

Client

YubiHSM client: main API in this crate for accessing functions of the HSM hardware device.

Enums

ErrorKind

Client error kinds

Type Aliases

Error

Client errors