Struct alloy::signers::local::yubihsm::domain::Domain

pub struct Domain(<Domain as PublicFlags>::Internal);

Available on crate features signers and signer-local only.

Logical partition within the YubiHSM 2, allowing several clients to access the same device but access controlled on a domain-by-domain basis. For more information, see the Yubico documentation:

https://developers.yubico.com/YubiHSM2/Concepts/Domain.html

Tuple Fields
0: <Domain as PublicFlags>::Internal
Implementations

impl Domain

pub const DOM1: Domain = _

pub const DOM2: Domain = _

pub const DOM3: Domain = _

pub const DOM4: Domain = _

pub const DOM5: Domain = _

pub const DOM6: Domain = _

pub const DOM7: Domain = _

pub const DOM8: Domain = _

pub const DOM9: Domain = _

pub const DOM10: Domain = _

pub const DOM11: Domain = _

pub const DOM12: Domain = _

pub const DOM13: Domain = _

pub const DOM14: Domain = _

pub const DOM15: Domain = _

pub const DOM16: Domain = _

impl Domain

pub const fn empty() -> Domain

Get a flags value with all bits unset.

pub const fn all() -> Domain

Get a flags value with all known bits set.

pub const fn bits(&self) -> u16

Get the underlying bits value.

The returned value is exactly the bits set in this flags value.

pub const fn from_bits(bits: u16) -> Option<Domain>

Convert from a bits value.

This method will return None if any unknown bits are set.

pub const fn from_bits_truncate(bits: u16) -> Domain

Convert from a bits value, unsetting any unknown bits.

pub const fn from_bits_retain(bits: u16) -> Domain

Convert from a bits value exactly.

pub fn from_name(name: &str) -> Option<Domain>

Get a flags value with the bits of a flag with the given name set.

This method will return None if name is empty or doesnโt correspond to any named flag.

pub const fn is_empty(&self) -> bool

Whether all bits in this flags value are unset.

pub const fn is_all(&self) -> bool

Whether all known bits in this flags value are set.

pub const fn intersects(&self, other: Domain) -> bool

Whether any set bits in a source flags value are also set in a target flags value.

pub const fn contains(&self, other: Domain) -> bool

Whether all set bits in a source flags value are also set in a target flags value.

pub fn insert(&mut self, other: Domain)

The bitwise or (|) of the bits in two flags values.

pub fn remove(&mut self, other: Domain)

The intersection of a source flags value with the complement of a target flags value (&!).

This method is not equivalent to self & !other when other has unknown bits set. remove wonโt truncate other, but the ! operator will.

pub fn toggle(&mut self, other: Domain)

The bitwise exclusive-or (^) of the bits in two flags values.

pub fn set(&mut self, other: Domain, value: bool)

Call insert when value is true or remove when value is false.

pub const fn intersection(self, other: Domain) -> Domain

The bitwise and (&) of the bits in two flags values.

pub const fn union(self, other: Domain) -> Domain

The bitwise or (|) of the bits in two flags values.

pub const fn difference(self, other: Domain) -> Domain

The intersection of a source flags value with the complement of a target flags value (&!).

This method is not equivalent to self & !other when other has unknown bits set. difference wonโt truncate other, but the ! operator will.

pub const fn symmetric_difference(self, other: Domain) -> Domain

The bitwise exclusive-or (^) of the bits in two flags values.

pub const fn complement(self) -> Domain

The bitwise negation (!) of the bits in a flags value, truncating the result.

impl Domain

pub const fn iter(&self) -> Iter<Domain>

Yield a set of contained flags values.

Each yielded flags value will correspond to a defined named flag. Any unknown bits will be yielded together as a final flags value.

pub const fn iter_names(&self) -> IterNames<Domain>

Yield a set of contained named flags values.

This method is like iter, except only yields bits in contained named flags. Any unknown bits, or bits not corresponding to a contained flag will not be yielded.

impl Domain

pub fn at(index: usize) -> Result<Domain, Error<ErrorKind>>

Get the Domain object corresponding to the given-numbered domain e.g. Domain::at(1) returns Domain::DOM1.