Module alloy::signers::local::yubihsm::object

Available on crate features signers and signer-local only.

Objects within the YubiHSM 2 (keys, certificates, or other opaque data)

For more information, see: https://developers.yubico.com/YubiHSM2/Concepts/Object.html

Modules

put

The YubiHSM 2 doesnt’ have a unified put object command, however all of the put object commands share a common structure, i.e. object::import::Params

Structs

Entry

Brief information about an object as included in ListObjectsCommand

Handle

Objects in the HSM are keyed by a tuple of their type an object::Id (i.e. multiple objects of different types can have the same object::Id)

Info

Information about an object

Label

Labels attached to objects

Enums

ErrorKind

Kinds of Object-related errors

Filter

Filters to apply when listing objects

Origin

Information about how a key was originally generated

Type

Types of objects

Constants

LABEL_SIZE

Number of bytes in a label on an object (fixed-size)

Type Aliases

Error

Object-related errors

Id

Object identifiers

SequenceId

Sequence identifiers: number of times an object with a given ID has been created in this YubiHSM (with all previous versions having been deleted)