Trait alloy::signers::SignerSync

pub trait SignerSync<Sig = Signature<Signature<Secp256k1>>> {
    // Required methods
    fn sign_hash_sync(&self, hash: &FixedBytes<32>) -> Result<Sig, Error>;
    fn chain_id_sync(&self) -> Option<u64>;

    // Provided methods
    fn sign_message_sync(&self, message: &[u8]) -> Result<Sig, Error> { ... }
    fn sign_typed_data_sync<T>(
        &self,
        payload: &T,
        domain: &Eip712Domain
    ) -> Result<Sig, Error>
       where T: SolStruct,
             Self: Sized { ... }
    fn sign_dynamic_typed_data_sync(
        &self,
        payload: &TypedData
    ) -> Result<Sig, Error> { ... }
}

Available on crate feature signers only.

Synchronous Ethereum signer.

All provided implementations rely on sign_hash_sync. A signer may not always be able to implement this method, in which case it should return UnsupportedOperation, and implement all the signing methods directly.

Synchronous signers should also implement Signer, as they are always able to by delegating the asynchronous methods to the synchronous ones.

Required Methods

fn sign_hash_sync(&self, hash: &FixedBytes<32>) -> Result<Sig, Error>

Signs the given hash.

fn chain_id_sync(&self) -> Option<u64>

Returns the signer’s chain ID.

Provided Methods

fn sign_message_sync(&self, message: &[u8]) -> Result<Sig, Error>

Signs the hash of the provided message after prefixing it, as specified in EIP-191.

fn sign_typed_data_sync<T>( &self, payload: &T, domain: &Eip712Domain ) -> Result<Sig, Error>
where T: SolStruct, Self: Sized,

Available on crate feature eip712 only.

Encodes and signs the typed data according to EIP-712.

fn sign_dynamic_typed_data_sync( &self, payload: &TypedData ) -> Result<Sig, Error>

Available on crate feature eip712 only.

Encodes and signs the typed data according to EIP-712 for Signers that are not dynamically sized.