Crate alloy::signers::wallet

Available on crate features signers and signer-wallet only.

alloy-signer-wallets

Local wallet implementations:

K256 private key

YubiHSM2

Features

keystore: enables Ethereum keystore functionality on the LocalWallet type.

mnemonic: enables BIP-39 mnemonic functionality for building LocalWallets.

yubihsm: enables Wallets with YubiHSM2 support.

Modules

coins_bip39

The bip39 crate is heavily inspired by and reuses code from Wagyu under the MIT license. The difference being, the underlying extended private keys are generated using the bip32 crate, that depends on k256 instead of libsecp256k1.

yubihsm

Structs

MnemonicBuilder

Represents a structure that can resolve into a LocalWallet.

Wallet

An Ethereum private-public key pair which can be used for signing messages.

Enums

WalletError

Error thrown by Wallet.

Type Aliases

LocalWallet

A wallet instantiated with a locally stored private key

YubiWalletyubihsm

A wallet instantiated with a YubiHSM