Trait alloy::signers::wallet::yubihsm::asymmetric::signature::digest::block_buffer::generic_array::sequence::Shorten

pub unsafe trait Shorten<T>: Sized + GenericSequence<T> {
    type Shorter: Lengthen<T, Longer = Self>;

    // Required methods
    fn pop_back(self) -> (Self::Shorter, T);
    fn pop_front(self) -> (T, Self::Shorter);
}

Available on crate features signers and signer-wallet only.

Defines a GenericSequence which can be shortened by removing the first or last element from it.

Additionally, any shortened sequence can be lengthened by appending or prepending an element to it.

Required Associated Types

type Shorter: Lengthen<T, Longer = Self>

GenericSequence that has one less element than Self

Required Methods

fn pop_back(self) -> (Self::Shorter, T)

Returns a new array without the last element, and the last element.

Example:

let a = arr![i32; 1, 2, 3, 4];

let (init, last) = a.pop_back();

assert_eq!(init, arr![i32; 1, 2, 3]);
assert_eq!(last, 4);

fn pop_front(self) -> (T, Self::Shorter)

Returns a new array without the first element, and the first element. Example:

let a = arr![i32; 1, 2, 3, 4];

let (head, tail) = a.pop_front();

assert_eq!(head, 1);
assert_eq!(tail, arr![i32; 2, 3, 4]);

Object Safety

This trait is not object safe.