Trait alloy::signers::wallet::yubihsm::asymmetric::signature::digest::block_buffer::generic_array::typenum::uint::SetBit

pub trait SetBit<I, B> {
    type Output;
}

Available on crate features signers and signer-wallet only.

A type operator that, when implemented for unsigned integer N, sets the bit at position I to B.

Required Associated Types

type Output

Object Safety

This trait is not object safe.