Trait alloy::signers::wallet::yubihsm::asymmetric::signature::PrehashSignature

pub trait PrehashSignature {
    type Digest: Digest;
}

Available on crate features signers and signer-wallet only.

Marker trait for Signature types computable as ğ’(ğ‡(ğ’)) i.e. ones which prehash a message to be signed as ğ‡(ğ’)

Where:

ğ’: signature algorithm

ğ‡: hash (a.k.a. digest) function

ğ’: message

This approach is relatively common in signature schemes based on the Fiat-Shamir heuristic.

For signature types that implement this trait, when the derive crate feature is enabled a custom derive for Signer is available for any types that impl DigestSigner, and likewise for deriving Verifier for types which impl DigestVerifier.

Required Associated Types

type Digest: Digest

Preferred Digest algorithm to use when computing this signature type.