Trait alloy::signers::wallet::yubihsm::asymmetric::signature::RandomizedSigner

pub trait RandomizedSigner<S> {
    // Required method
    fn try_sign_with_rng(
        &self,
        rng: &mut impl CryptoRngCore,
        msg: &[u8]
    ) -> Result<S, Error>;

    // Provided method
    fn sign_with_rng(&self, rng: &mut impl CryptoRngCore, msg: &[u8]) -> S { ... }
}

Available on crate features signers and signer-wallet and rand_core only.

Sign the given message using the provided external randomness source.

Required Methods

fn try_sign_with_rng( &self, rng: &mut impl CryptoRngCore, msg: &[u8] ) -> Result<S, Error>

Attempt to sign the given message, returning a digital signature on success, or an error if something went wrong.

The main intended use case for signing errors is when communicating with external signers, e.g. cloud KMS, HSMs, or other hardware tokens.

Provided Methods

fn sign_with_rng(&self, rng: &mut impl CryptoRngCore, msg: &[u8]) -> S

Sign the given message and return a digital signature

Object Safety

This trait is not object safe.