Module alloy::signers::wallet::yubihsm::authentication

Available on crate features signers and signer-wallet only.

YubiHSM 2 authentication functionality (i.e. credentials used to authenticate and establish a session with an HSM)

Modules

commands

Put an existing auth key into the YubiHSM 2

key

YubiHSM 2 authentication keys (2 * AES-128 symmetric PSK) from which session keys are derived

Structs

Credentials

Credentials used to establish a session with the HSM

Key

YubiHSM 2 authentication keys (2 * AES-128 symmetric PSK) from which session keys are derived.c

Enums

Algorithm

Valid algorithms for auth keys

ErrorKind

Kinds of authentication errors

Constants

DEFAULT_AUTHENTICATION_KEY_ID

Default auth key ID slot

Type Aliases

Error

Authentication errors