Module alloy::signers::wallet::yubihsm::authentication::key

Available on crate features signers and signer-wallet only.

YubiHSM 2 authentication keys (2 * AES-128 symmetric PSK) from which session keys are derived

Structs

Key

YubiHSM 2 authentication keys (2 * AES-128 symmetric PSK) from which session keys are derived.c

Constants

DEFAULT_PASSWORD

Password from which the default auth key is derived

PBKDF2_ITERATIONS

Number of PBKDF2 iterations to perform when deriving auth keys. This number of iterations matches what is performed by yubihsm-shell.

PBKDF2_SALT

Salt value to use with PBKDF2 when deriving auth keys from a password. This salt is designed to be compatible with the password functionality in yubihsm-shell (otherwise using a static salt is not best practice).

SIZE

Auth keys are 2 * AES-128 keys