Struct alloy::signers::wallet::yubihsm::client::Client

pub struct Client {
    connector: Connector,
    session: Arc<Mutex<Option<Session>>>,
    credentials: Option<Credentials>,
}

Available on crate features signers and signer-wallet only.

YubiHSM client: main API in this crate for accessing functions of the HSM hardware device.

Fields
connector: Connectorsession: Arc<Mutex<Option<Session>>>credentials: Option<Credentials>
Implementations

impl Client

pub fn open( connector: Connector, credentials: Credentials, reconnect: bool ) -> Result<Client, Error<ErrorKind>>

Open a connection via a Connector to a YubiHSM, returning a yubihsm::Client. Valid Connector types are: HttpConnector, UsbConnector, and MockHsm.

pub fn create( connector: Connector, credentials: Credentials ) -> Result<Client, Error<ErrorKind>>

Create a yubihsm::Client, but defer connecting until connect() is called.

pub fn connector(&self) -> &Connector

Borrow this client’s YubiHSM connector (which is Cloneable)

pub fn connect(&self) -> Result<(), Error<ErrorKind>>

Connect to the HSM (idempotently, i.e. returns success if we have an open connection already)

pub fn session(&self) -> Result<Guard<'_>, Error<ErrorKind>>

Get current Session (either opening a new one or returning an already open one).

pub fn ping(&self) -> Result<Duration, Error<ErrorKind>>

Ping the HSM, ensuring we have a live connection and returning the end-to-end latency.

pub fn blink_device(&self, num_seconds: u8) -> Result<(), Error<ErrorKind>>

Blink the HSM’s LEDs (to identify it) for the given number of seconds.

https://developers.yubico.com/YubiHSM2/Commands/Blink_Device.html

pub fn decrypt_oaep<T>( &self, key_id: u16, mgf1_hash_alg: Algorithm, data: T, label_hash: Vec<u8> ) -> Result<DecryptedData, Error<ErrorKind>>
where T: Into<Vec<u8>>,

Decrypt data encrypted with RSA-OAEP

https://developers.yubico.com/YubiHSM2/Commands/Decrypt_Oaep.html

pub fn delete_object( &self, object_id: u16, object_type: Type ) -> Result<(), Error<ErrorKind>>

Delete an object of the given ID and type.

https://developers.yubico.com/YubiHSM2/Commands/Delete_Object.html

pub fn device_info(&self) -> Result<Info, Error<ErrorKind>>

Get information about the HSM device.

https://developers.yubico.com/YubiHSM2/Commands/Device_Info.html

pub fn echo<M>(&self, msg: M) -> Result<Vec<u8>, Error<ErrorKind>>
where M: Into<Vec<u8>>,

Echo a message sent to the HSM.

https://developers.yubico.com/YubiHSM2/Commands/Echo.html

pub fn export_wrapped( &self, wrap_key_id: u16, object_type: Type, object_id: u16 ) -> Result<Message, Error<ErrorKind>>

Export an encrypted object from the HSM using the given key-wrapping key.

https://developers.yubico.com/YubiHSM2/Commands/Export_Wrapped.html

pub fn generate_asymmetric_key( &self, key_id: u16, label: Label, domains: Domain, capabilities: Capability, algorithm: Algorithm ) -> Result<u16, Error<ErrorKind>>

Generate a new asymmetric key within the HSM.

https://developers.yubico.com/YubiHSM2/Commands/Generate_Asymmetric_Key.html

pub fn generate_hmac_key( &self, key_id: u16, label: Label, domains: Domain, capabilities: Capability, algorithm: Algorithm ) -> Result<u16, Error<ErrorKind>>

Generate a new HMAC key within the HSM.

https://developers.yubico.com/YubiHSM2/Commands/Generate_Hmac_Key.html

pub fn generate_wrap_key( &self, key_id: u16, label: Label, domains: Domain, capabilities: Capability, delegated_capabilities: Capability, algorithm: Algorithm ) -> Result<u16, Error<ErrorKind>>

Generate a new wrap key within the HSM.

Delegated capabilities are the set of Capability bits that an object is allowed to have when imported or exported using the wrap key.

https://developers.yubico.com/YubiHSM2/Commands/Generate_Wrap_Key.html

pub fn get_log_entries(&self) -> Result<LogEntries, Error<ErrorKind>>

Get audit logs from the HSM device.

https://developers.yubico.com/YubiHSM2/Commands/Get_Log_Entries.html

pub fn get_object_info( &self, object_id: u16, object_type: Type ) -> Result<Info, Error<ErrorKind>>

Get information about an object.

https://developers.yubico.com/YubiHSM2/Commands/Get_Object_Info.html

pub fn get_opaque(&self, object_id: u16) -> Result<Vec<u8>, Error<ErrorKind>>

Get an opaque object stored in the HSM.

https://developers.yubico.com/YubiHSM2/Commands/Get_Opaque.html

pub fn get_command_audit_option( &self, command: Code ) -> Result<AuditOption, Error<ErrorKind>>

Get the audit policy setting for a particular command.

https://developers.yubico.com/YubiHSM2/Commands/Get_Option.html

pub fn get_commands_audit_options( &self ) -> Result<Vec<AuditCommand>, Error<ErrorKind>>

Get the audit policy settings for all commands.

https://developers.yubico.com/YubiHSM2/Commands/Get_Option.html

pub fn get_force_audit_option(&self) -> Result<AuditOption, Error<ErrorKind>>

Get the forced auditing global option: when enabled, the device will refuse operations if the log store becomes full.

https://developers.yubico.com/YubiHSM2/Commands/Get_Option.html

pub fn get_pseudo_random( &self, bytes: usize ) -> Result<Vec<u8>, Error<ErrorKind>>

Get some number of bytes of pseudo random data generated on the device.

https://developers.yubico.com/YubiHSM2/Commands/Get_Pseudo_Random.html

pub fn get_public_key(&self, key_id: u16) -> Result<PublicKey, Error<ErrorKind>>

Get the public key for an asymmetric key stored on the device.

https://developers.yubico.com/YubiHSM2/Commands/Get_Public_Key.html

pub fn get_storage_info(&self) -> Result<Info, Error<ErrorKind>>

Get storage info (i.e. currently free storage) from the HSM device.

https://developers.yubico.com/YubiHSM2/Commands/Get_Storage_Info.html

pub fn get_template(&self, object_id: u16) -> Result<Vec<u8>, Error<ErrorKind>>

Get a certificate template (i.e. for SSH CA) stored in the HSM.

https://developers.yubico.com/YubiHSM2/Commands/Get_Template.html

pub fn import_wrapped<M>( &self, wrap_key_id: u16, wrap_message: M ) -> Result<Handle, Error<ErrorKind>>
where M: Into<Message>,

Import an encrypted object from the HSM using the given key-wrapping key.

https://developers.yubico.com/YubiHSM2/Commands/Import_Wrapped.html

pub fn list_objects( &self, filters: &[Filter] ) -> Result<Vec<Entry>, Error<ErrorKind>>

List objects visible from the current session.

Optionally apply a set of provided filters which select objects based on their attributes.

https://developers.yubico.com/YubiHSM2/Commands/List_Objects.html

pub fn put_asymmetric_key<K>( &self, key_id: u16, label: Label, domains: Domain, capabilities: Capability, algorithm: Algorithm, key_bytes: K ) -> Result<u16, Error<ErrorKind>>
where K: Into<Vec<u8>>,

Put an existing asymmetric key into the HSM.

https://developers.yubico.com/YubiHSM2/Commands/Put_Asymmetric.html

pub fn put_authentication_key<K>( &self, key_id: u16, label: Label, domains: Domain, capabilities: Capability, delegated_capabilities: Capability, algorithm: Algorithm, authentication_key: K ) -> Result<u16, Error<ErrorKind>>
where K: Into<Key>,

Put an existing authentication::Key into the HSM.

https://developers.yubico.com/YubiHSM2/Commands/Put_Authentication_Key.html

pub fn put_hmac_key<K>( &self, key_id: u16, label: Label, domains: Domain, capabilities: Capability, algorithm: Algorithm, key_bytes: K ) -> Result<u16, Error<ErrorKind>>
where K: Into<Vec<u8>>,

Put an existing HMAC key into the HSM.

https://developers.yubico.com/YubiHSM2/Commands/Put_Hmac_Key.html

pub fn put_opaque<B>( &self, object_id: u16, label: Label, domains: Domain, capabilities: Capability, algorithm: Algorithm, opaque_data: B ) -> Result<u16, Error<ErrorKind>>
where B: Into<Vec<u8>>,

Put an opaque object (X.509 certificate or other bytestring) into the HSM.

https://developers.yubico.com/YubiHSM2/Commands/Put_Opaque.html

pub fn put_otp_aead_key<K>( &self, key_id: u16, label: Label, domains: Domain, capabilities: Capability, algorithm: Algorithm, key_bytes: K ) -> Result<u16, Error<ErrorKind>>
where K: Into<Vec<u8>>,

Put an existing OTP AEAD key into the HSM.

https://developers.yubico.com/YubiHSM2/Commands/Put_Otp_Aead_Key.html

pub fn put_wrap_key<K>( &self, key_id: u16, label: Label, domains: Domain, capabilities: Capability, delegated_capabilities: Capability, algorithm: Algorithm, key_bytes: K ) -> Result<u16, Error<ErrorKind>>
where K: Into<Vec<u8>>,

Put an existing wrap key into the HSM.

https://developers.yubico.com/YubiHSM2/Commands/Put_Wrap_Key.html

pub fn put_template<T>( &self, object_id: u16, label: Label, domains: Domain, capabilities: Capability, template: T ) -> Result<u16, Error<ErrorKind>>
where T: Into<Template>,

Put a template object (i.e. for SSH CA) into the HSM.

Use the yubihsm::ssh::Template type for SSH CA templates.

https://developers.yubico.com/YubiHSM2/Commands/Put_Template.html

pub fn reset_device(&self) -> Result<(), Error<ErrorKind>>

Reset the HSM to a factory default state and reboot, clearing all stored objects and restoring the default auth key.

WARNING: This wipes all keys and other data from the HSM! Make absolutely sure you want to use this!

https://developers.yubico.com/YubiHSM2/Commands/Reset_Device.html

pub fn reset_device_and_reconnect( &mut self, timeout: Duration ) -> Result<(), Error<ErrorKind>>

Available on crate feature passwords only.

Reset the HSM to a factory default state and reboot, clearing all stored objects and restoring the default auth key. This method further attempts to wait for the HSM to finish resetting and then attempts to reauthenticate with the default credentials.

Upon successfully resetting the device and autenticating using the default administrator credentials in key slot 0x01, a new yubihsm::Client is returned.

WARNING: This wipes all keys and other data from the HSM! Make absolutely sure you want to use this!

https://developers.yubico.com/YubiHSM2/Commands/Reset_Device.html

pub fn set_command_audit_option( &self, command: Code, audit_option: AuditOption ) -> Result<(), Error<ErrorKind>>

Configure the audit policy settings for a particular command, e.g. auditing should be On, Off, or Fix (i.e. fixed permanently on).

https://developers.yubico.com/YubiHSM2/Commands/Set_Option.html

pub fn set_force_audit_option( &self, option: AuditOption ) -> Result<(), Error<ErrorKind>>

Put the forced auditing global option: when enabled, the device will refuse operations if the log store becomes full.

Options are On, Off, or Fix (i.e. fixed permanently on)

https://developers.yubico.com/YubiHSM2/Commands/Put_Option.html

pub fn set_log_index(&self, log_index: u16) -> Result<(), Error<ErrorKind>>

Set the index of the last consumed index of the HSM audit log.

https://developers.yubico.com/YubiHSM2/Commands/Set_Log_Index.html

pub fn sign_attestation_certificate( &self, key_id: u16, attestation_key_id: Option<u16> ) -> Result<Certificate, Error<ErrorKind>>

Obtain an X.509 attestation certificate for a key within the HSM. This can be used to demonstrate that a given key was generated by and stored within a HSM in a non-exportable manner.

The key_id is the subject key for which an attestation certificate is created, and theattestation_key_id will be used to sign the attestation certificate.

If no attestation key is given, the device’s default attestation key will be used, and can be verified against Yubico’s certificate.

https://developers.yubico.com/YubiHSM2/Commands/Sign_Attestation_Certificate.html

pub fn sign_ecdsa_prehash_raw<T>( &self, key_id: u16, digest: T ) -> Result<Vec<u8>, Error<ErrorKind>>
where T: Into<Vec<u8>>,

Compute an ECDSA signature of the given digest (i.e. a precomputed SHA-2 digest)

https://developers.yubico.com/YubiHSM2/Commands/Sign_Ecdsa.html

Security Warning

This is a low-level ECDSA API, and if used incorrectly could potentially result in forgeable signatures.

We recommend using the [ecdsa::Signer] type instead, which provides a high-level, well-typed, misuse resistant API.

pub fn sign_ed25519<T>( &self, key_id: u16, data: T ) -> Result<Signature, Error<ErrorKind>>
where T: Into<Vec<u8>>,

Compute an Ed25519 signature with the given key ID.

https://developers.yubico.com/YubiHSM2/Commands/Sign_Eddsa.html

pub fn sign_hmac<M>(&self, key_id: u16, msg: M) -> Result<Tag, Error<ErrorKind>>
where M: Into<Vec<u8>>,

Compute an HMAC tag of the given data with the given key ID.

https://developers.yubico.com/YubiHSM2/Commands/Sign_Hmac.html

pub fn unwrap_data<M>( &self, wrap_key_id: u16, wrap_message: M ) -> Result<Vec<u8>, Error<ErrorKind>>
where M: Into<Message>,

Decrypt data which was encrypted (using AES-CCM) under a wrap key.

https://developers.yubico.com/YubiHSM2/Commands/Unwrap_Data.html

pub fn verify_hmac<M, T>( &self, key_id: u16, msg: M, tag: T ) -> Result<(), Error<ErrorKind>>
where M: Into<Vec<u8>>, T: Into<Tag>,

Verify an HMAC tag of the given data with the given key ID.

https://developers.yubico.com/YubiHSM2/Commands/Verify_Hmac.html

pub fn wrap_data( &self, wrap_key_id: u16, plaintext: Vec<u8> ) -> Result<Message, Error<ErrorKind>>

Encrypt data (with AES-CCM) using the given wrap key.

https://developers.yubico.com/YubiHSM2/Commands/Wrap_Data.html