Module alloy::signers::wallet::yubihsm::connector::usb

Available on crate features signers and signer-wallet and usb only.

Support for connecting to the YubiHSM 2 via USB.

Typically to access a YubiHSM 2 via USB, you’ll use the Connector::usb method in the event there is only one expected to be connected at a time.

To enumerate available USB devices (e.g. in the case there is more than one YubiHSM connected to the same computer), use Devices.

Structs

Device

A USB device we’ve identified as a YubiHSM 2

Devices

A collection of detected YubiHSM 2 devices, represented as Device

UsbConfig

Configuration for connecting to the YubiHSM via USB

UsbConnection

Connection to HSM via USB

UsbConnector

Connect to the HSM via USB.

UsbTimeout

Timeouts when performing USB operations

Constants

YUBICO_VENDOR_ID

USB vendor ID for Yubico

YUBIHSM2_BULK_IN_ENDPOINT

YubiHSM 2 bulk in endpoint

YUBIHSM2_BULK_OUT_ENDPOINT

YubiHSM 2 bulk out endpoint

YUBIHSM2_INTERFACE_NUM

YubiHSM 2 USB interface number

YUBIHSM2_PRODUCT_ID

USB product ID for the YubiHSM 2