Module alloy::signers::wallet::yubihsm::ecdsa::sec1

Available on crate features signers and signer-wallet and sec1 only.

Support for SEC1 elliptic curve encoding formats.

https://www.secg.org/sec1-v2.pdf

Enums

Coordinates

Enum representing the coordinates of either compressed or uncompressed SEC1-encoded elliptic curve points.

Tag

Tag byte used by the Elliptic-Curve-Point-to-Octet-String encoding.

Traits

FromEncodedPoint

Trait for deserializing a value from a SEC1 encoded curve point.

ModulusSize

Trait for supported modulus sizes which precomputes the typenums for various point encodings so they don’t need to be included as bounds.

ToCompactEncodedPoint

Trait for serializing a value to a SEC1 encoded curve point with compaction.

ToEncodedPoint

Trait for serializing a value to a SEC1 encoded curve point.

ValidatePublicKey

Validate that the given EncodedPoint represents the encoded public key value of the given secret.

Type Aliases

CompressedPoint

Encoded elliptic curve point with point compression.

CompressedPointSize

Size of a compressed elliptic curve point.

EncodedPoint

Encoded elliptic curve point sized appropriately for a given curve.

UncompressedPoint

Encoded elliptic curve point without point compression.

UncompressedPointSize

Size of an uncompressed elliptic curve point.