Module alloy::signers::wallet::yubihsm::rsa

Available on crate features signers and signer-wallet only.

RSA (Rivest–Shamir–Adleman) asymmetric cryptosystem support (signing/encryption).

NOTE: This functionality has not been properly tested and is therefore not enabled by default! Enable the untested cargo feature if you would like to use it (please report success or bugs!)

Modules

mgf

Mask generating functions for use with RSASSA-PSS signatures

oaep

RSA encryption with Optimal Asymmetric Encryption Padding (OAEP)

pkcs1

RSASSA-PKCS#1v1.5 commands

pss

RSASSA-PSS: Probabilistic Signature Scheme based on the RSASP1 and RSAVP1 primitives with the EMSA-PSS encoding method.

Enums

Algorithm

RSA algorithms (signing and encryption)