Module alloy::transports::http::hyper::body

Available on crate features transports and transport-http only.

Streaming bodies for Requests and Responses

For both Clients and Servers, requests and responses use streaming bodies, instead of complete buffering. This allows applications to not use memory they don’t need, and allows exerting back-pressure on connections by only reading when asked.

There are two pieces to this in hyper:

The Body trait describes all possible bodies. hyper allows any body type that implements Body, allowing applications to have fine-grained control over their streaming.

The Incoming concrete type, which is an implementation of Body, and returned by hyper as a “receive stream” (so, for server requests and client responses).

There are additional implementations available in http-body-util, such as a Full or Empty body.

Structs

Bytes

A cheaply cloneable and sliceable chunk of contiguous memory.

Frame

A frame of any kind related to an HTTP stream (body).

Incoming

A stream of Bytes, used when receiving bodies from the network.

SizeHint

A Body size hint

Traits

Body

Trait representing a streaming body of a Request or Response.

Buf

Read bytes from a buffer.