Trait alloy::transports::http::hyper::rt::bounds::Http2ClientConnExec

pub trait Http2ClientConnExec<B, T>: Sealed<(B, T)>
where
    B: Body,
    <B as Body>::Error: Into<Box<dyn Error + Send + Sync>>,
    T: Read + Write + Unpin,
{ }

Available on crate features transports and transport-http only.

An executor to spawn http2 futures for the client.

This trait is implemented for any type that implements Executor trait for any future.

This trait is sealed and cannot be implemented for types outside this crate.