Trait alloy::transports::http::hyper_util::client::legacy::connect::Connect

pub trait Connect: Sized + Sealed { }

Available on crate features transports and transport-http and client and client-legacy only.

Connect to a destination, returning an IO transport.

A connector receives a Uri and returns a Future of the ready connection.

Trait Alias

This is really just an alias for the tower::Service trait, with additional bounds set for convenience inside hyper. You don’t actually implement this trait, but tower::Service<Uri> instead.

Object Safety

This trait is not object safe.