Trait alloy::transports::http::hyper_util::server::graceful::GracefulConnection

pub trait GracefulConnection: Future<Output = Result<(), Self::Error>> + Sealed {
    type Error;

    // Required method
    fn graceful_shutdown(self: Pin<&mut Self>);
}

Available on crate features transports and transport-http and server and server-graceful only.

An internal utility trait as an umbrella target for all (hyper) connection types that the GracefulShutdown can watch.

Required Associated Types

type Error

The error type returned by the connection when used as a future.

Required Methods

fn graceful_shutdown(self: Pin<&mut Self>)

Start a graceful shutdown process for this connection.