alloy::transports::http::reqwest::blocking

Struct Response

pub struct Response { /* private fields */ }

A Response to a submitted Request.

Implementations

impl Response

pub fn status(&self) -> StatusCode

Get the StatusCode of this Response.

Examples

Checking for general status class:

let resp = reqwest::blocking::get("http://httpbin.org/get")?;
if resp.status().is_success() {
    println!("success!");
} else if resp.status().is_server_error() {
    println!("server error!");
} else {
    println!("Something else happened. Status: {:?}", resp.status());
}

Checking for specific status codes:

use reqwest::blocking::Client;
use reqwest::StatusCode;
let client = Client::new();

let resp = client.post("http://httpbin.org/post")
    .body("possibly too large")
    .send()?;

match resp.status() {
    StatusCode::OK => println!("success!"),
    StatusCode::PAYLOAD_TOO_LARGE => {
        println!("Request payload is too large!");
    }
    s => println!("Received response status: {s:?}"),
};

pub fn headers(&self) -> &HeaderMap

Get the Headers of this Response.

Example

Saving an etag when caching a file:

use reqwest::blocking::Client;
use reqwest::header::ETAG;

let client = Client::new();

let mut resp = client.get("http://httpbin.org/cache").send()?;
if resp.status().is_success() {
    if let Some(etag) = resp.headers().get(ETAG) {
        std::fs::write("etag", etag.as_bytes());
    }
    let mut file = std::fs::File::create("file")?;
    resp.copy_to(&mut file)?;
}

pub fn headers_mut(&mut self) -> &mut HeaderMap

Get a mutable reference to the Headers of this Response.

pub fn version(&self) -> Version

Get the HTTP Version of this Response.

pub fn url(&self) -> &Url

Get the final Url of this Response.

Example

let resp = reqwest::blocking::get("http://httpbin.org/redirect/1")?;
assert_eq!(resp.url().as_str(), "http://httpbin.org/get");

pub fn remote_addr(&self) -> Option<SocketAddr>

Get the remote address used to get this Response.

Example

let resp = reqwest::blocking::get("http://httpbin.org/redirect/1")?;
println!("httpbin.org address: {:?}", resp.remote_addr());

pub fn extensions(&self) -> &Extensions

Returns a reference to the associated extensions.

pub fn extensions_mut(&mut self) -> &mut Extensions

Returns a mutable reference to the associated extensions.

pub fn content_length(&self) -> Option<u64>

Get the content length of the response, if it is known.

This value does not directly represents the value of the Content-Length header, but rather the size of the responseโs body. To read the headerโs value, please use the Response::headers method instead.

Reasons it may not be known:

The response does not include a body (e.g. it responds to a HEAD request).

The response is gzipped and automatically decoded (thus changing the actual decoded length).

pub fn json<T>(self) -> Result<T, Error>
where T: DeserializeOwned,

Try and deserialize the response body as JSON using serde.

Optional

This requires the optional json feature enabled.

Examples

// This `derive` requires the `serde` dependency.
#[derive(Deserialize)]
struct Ip {
    origin: String,
}

let json: Ip = reqwest::blocking::get("http://httpbin.org/ip")?.json()?;

Errors

This method fails whenever the response body is not in JSON format, or it cannot be properly deserialized to target type T. For more details please see serde_json::from_reader.

pub fn bytes(self) -> Result<Bytes, Error>

Get the full response body as Bytes.

Example

let bytes = reqwest::blocking::get("http://httpbin.org/ip")?.bytes()?;

println!("bytes: {bytes:?}");

pub fn text(self) -> Result<String, Error>

Get the response text.

This method decodes the response body with BOM sniffing and with malformed sequences replaced with the char::REPLACEMENT_CHARACTER. Encoding is determined from the charset parameter of Content-Type header, and defaults to utf-8 if not presented.

Note

If the charset feature is disabled the method will only attempt to decode the response as UTF-8, regardless of the given Content-Type

Example

let content = reqwest::blocking::get("http://httpbin.org/range/26")?.text()?;

pub fn copy_to<W>(&mut self, w: &mut W) -> Result<u64, Error>
where W: Write + ?Sized,

Copy the response body into a writer.

This function internally uses std::io::copy and hence will continuously read data from the body and then write it into writer in a streaming fashion until EOF is met.

On success, the total number of bytes that were copied to writer is returned.

Example

let mut resp = reqwest::blocking::get("http://httpbin.org/range/5")?;
let mut buf: Vec<u8> = vec![];
resp.copy_to(&mut buf)?;
assert_eq!(b"abcde", buf.as_slice());

pub fn error_for_status(self) -> Result<Response, Error>

Turn a response into an error if the server returned an error.

Example

let res = reqwest::blocking::get("http://httpbin.org/status/400")?
    .error_for_status();
if let Err(err) = res {
    assert_eq!(err.status(), Some(reqwest::StatusCode::BAD_REQUEST));
}

pub fn error_for_status_ref(&self) -> Result<&Response, Error>

Turn a reference to a response into an error if the server returned an error.

Example

let res = reqwest::blocking::get("http://httpbin.org/status/400")?;
let res = res.error_for_status_ref();
if let Err(err) = res {
    assert_eq!(err.status(), Some(reqwest::StatusCode::BAD_REQUEST));
}