alloy::transports::http::reqwest

Trait ResponseBuilderExt

pub trait ResponseBuilderExt {
    // Required method
    fn url(self, url: Url) -> Self;
}

Extension trait for http::response::Builder objects

Allows the user to add a Url to the http::Response

Required Methods

fn url(self, url: Url) -> Self

A builder method for the http::response::Builder type that allows the user to add a Url to the http::Response

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.