alloy::transports

Trait DualTransportHandler

pub trait DualTransportHandler<L, R> {
    // Required method
    fn call(
        &self,
        request: RequestPacket,
        left: L,
        right: R,
    ) -> Pin<Box<dyn Future<Output = Result<ResponsePacket, RpcError<TransportErrorKind>>> + Send>>;
}

Trait that determines how to dispatch a request given two transports.

Required Methods

fn call( &self, request: RequestPacket, left: L, right: R, ) -> Pin<Box<dyn Future<Output = Result<ResponsePacket, RpcError<TransportErrorKind>>> + Send>>

The type of the future returned by the transport.