alloy::transports

Trait IntoBoxTransport

pub trait IntoBoxTransport:
    Transport
    + Clone
    + Sealed {
    // Required method
    fn into_box_transport(self) -> BoxTransport;
}

Trait for converting a transport into a boxed transport.

This trait is sealed and implemented for all types that implement Transport + Clone.

Required Methods

fn into_box_transport(self) -> BoxTransport

Boxes the transport.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.