Crate alloy

Alloy

Alloy connects applications to blockchains.

Alloy is a rewrite of ethers-rs from the ground up, with exciting new features, high performance, and excellent docs.

We also have a book on all things Alloy and many examples to help you get started.

Installation

Alloy consists of a number of crates that provide a range of functionality essential for interfacing with any Ethereum-based blockchain.

The easiest way to get started is to add the alloy crate with the full feature flag from the command-line using Cargo:

cargo add alloy --features full

Alternatively, you can add the following to your Cargo.toml file:

alloy = { version = "0.3", features = ["full"] }

For a more fine-grained control over the features you wish to include, you can add the individual crates to your Cargo.toml file, or use the alloy crate with the features you need.

A comprehensive list of available features can be found on docs.rs or in the alloy crate’s Cargo.toml.

Overview

This repository contains the following crates:

alloy: Meta-crate for the entire project, including alloy-core

alloy-consensus - Ethereum consensus interface

alloy-consensus-any - Catch-all consensus interface for multiple networks

alloy-contract - Interact with on-chain contracts

alloy-eips - Ethereum Improvement Proposal (EIP) implementations

alloy-genesis - Ethereum genesis file definitions

alloy-json-rpc - Core data types for JSON-RPC 2.0 clients

alloy-network - Network abstraction for RPC types

alloy-network-primitives - Primitive types for the network abstraction

alloy-node-bindings - Ethereum execution-layer client bindings

alloy-provider - Interface with an Ethereum blockchain

alloy-pubsub - Ethereum JSON-RPC publish-subscribe tower service and type definitions

alloy-rpc-client - Low-level Ethereum JSON-RPC client implementation

alloy-rpc-types - Meta-crate for all Ethereum JSON-RPC types

alloy-rpc-types-admin - Types for the admin Ethereum JSON-RPC namespace

alloy-rpc-types-anvil - Types for the Anvil development node’s Ethereum JSON-RPC namespace

alloy-rpc-types-any - Types for JSON-RPC namespaces across multiple networks

alloy-rpc-types-beacon - Types for the Ethereum Beacon Node API

alloy-rpc-types-debug - Types for the debug Ethereum JSON-RPC namespace

alloy-rpc-types-engine - Types for the engine Ethereum JSON-RPC namespace

alloy-rpc-types-eth - Types for the eth Ethereum JSON-RPC namespace

alloy-rpc-types-mev - Types for the MEV bundle JSON-RPC namespace

alloy-rpc-types-trace - Types for the trace Ethereum JSON-RPC namespace

alloy-rpc-types-txpool - Types for the txpool Ethereum JSON-RPC namespace

alloy-serde - Serde-related utilities

alloy-signer - Ethereum signer abstraction

alloy-signer-aws - AWS KMS signer implementation

alloy-signer-gcp - GCP KMS signer implementation

alloy-signer-ledger - Ledger signer implementation

alloy-signer-local - Local (private key, keystore, mnemonic, YubiHSM) signer implementations

alloy-signer-trezor - Trezor signer implementation

alloy-transport - Low-level Ethereum JSON-RPC transport abstraction

alloy-transport-http - HTTP transport implementation

alloy-transport-ipc - IPC transport implementation

alloy-transport-ws - WS transport implementation

Supported Rust Versions (MSRV)

The current MSRV (minimum supported rust version) is 1.81.

Alloy will keep a rolling MSRV policy of at least two versions behind the latest stable release (so if the latest stable release is 1.58, we would support 1.56).

Note that the MSRV is not increased automatically, and only as part of a patch (pre-1.0) or minor (post-1.0) release.

Contributing

Thanks for your help improving the project! We are so happy to have you! We have a contributing guide to help you get involved in the Alloy project.

Pull requests will not be merged unless CI passes, so please ensure that your contribution follows the linting rules and passes clippy.

Note on no_std

Because these crates are primarily network-focused, we do not intend to support no_std for most of them at this time.

The following crates support no_std:

alloy-eips

alloy-genesis

alloy-serde

alloy-consensus

If you would like to add no_std support to a crate, please make sure to update scripts/check_no_std.sh as well.

Credits

None of these crates would have been possible without the great work done in:

ethers.js

rust-web3

ruint

ethabi

ethcontract-rs

guac_rs

License
Licensed under either of Apache License, Version 2.0 or MIT license at your option.
Unless you explicitly state otherwise, any contribution intentionally submitted for inclusion in these crates by you, as defined in the Apache-2.0 license, shall be dual licensed as above, without any additional terms or conditions.

Re-exports

pub use primitives::hex;

pub use primitives::hex;

pub use primitives::uint;

pub use sol_types::sol;

Modules

consensus

alloy-consensus

contract

alloy-contract

dyn_abi

alloy-dyn-abi

eips

alloy-eips

json_abi

alloy-json-abi

network

alloy-networks

node_bindings

alloy-node-bindings

primitives

alloy-primitives

providers

Interface with an Ethereum blockchain.

rpc

Ethereum JSON-RPC client and types.

signers

Ethereum signer abstraction and implementations.

sol_types

alloy-sol-types

transports

Low-level Ethereum JSON-RPC transport abstraction and implementations.