Crate alloy::consensus

alloy-consensus

Ethereum consensus interface.

This crate contains constants, types, and functions for implementing Ethereum EL consensus and communication. This includes headers, blocks, transactions, EIP-2718 envelopes, EIP-2930, EIP-4844, and more.

In general a type belongs in this crate if it is committed to in the EL block header. This includes:

transactions

blocks

headers

receipts

EIP-2718 envelopes.

Provenance

Much of this code was ported from reth-primitives as part of ongoing alloy migrations.

Modules

conditional

Helpers for conditional transactions.

constants

Ethereum protocol-related constants

proofs

Helper function for calculating Merkle proofs and hashes.

transaction

Transaction types.

Structs

Account

Represents an TrieAccount in the account trie.

BlobTransactionSidecar

This represents a set of blobs, and its corresponding commitments and proofs.

Block

Ethereum full block.

BlockBody

A response to GetBlockBodies, containing bodies if any bodies were found.

Header

Ethereum Block header

Receipt

Receipt containing result of transaction execution.

ReceiptWithBloom

Receipt with calculated bloom filter.

Receipts

Receipt containing result of transaction execution.

Sealed

A consensus hashable item, with its memoized hash.

SidecarBuilder

Build a BlobTransactionSidecar from an arbitrary amount of data.

Signed

A transaction with a signature and hash seal.

SimpleCoder

Simple coder that only uses the last 31 bytes of each blob. This is the default coder for the SidecarBuilder.

TrieAccount

Represents an TrieAccount in the account trie.

TxEip1559

A transaction with a priority fee (EIP-1559).

TxEip2930

Transaction with an AccessList (EIP-2930).

TxEip4844

EIP-4844 Blob Transaction

TxEip7702

A transaction with a priority fee (EIP-7702).

TxEip4844WithSidecar

EIP-4844 Blob Transaction

TxLegacy

Legacy transaction.

Enums

Eip658Value

Captures the result of a transaction execution.

ReceiptEnvelope

Receipt envelope, as defined in EIP-2718.

TxEip4844Variant

EIP-4844 Blob Transaction

TxEnvelope

The Ethereum EIP-2718 Transaction Envelope.

TxType

Ethereum TransactionType flags as specified in EIPs 2718, 1559, 2930, 4844, and 7702.

TypedTransaction

The TypedTransaction enum represents all Ethereum transaction request types.

Constants

EMPTY_OMMER_ROOT_HASH

Ommer root of empty list.

EMPTY_ROOT_HASH

Root hash of an empty trie.

Traits

BlockHeader

Trait for extracting specific Ethereum block data from a header

Eip2718EncodableReceipt

Receipt type that knows its EIP-2718 encoding.

RlpDecodableReceipt

Receipt type that knows how to decode itself with a Bloom value.

RlpEncodableReceipt

Receipt type that knows how to encode itself with a Bloom value.

Sealable

Sealeable objects.

SidecarCoder

A strategy for coding and decoding data into sidecars.

SignableTransaction

A signable transaction.

Transaction

Represents a minimal EVM transaction.

TxReceipt

Receipt is the result of a transaction execution.

Typed2718

A trait that helps to determine the type of the transaction.

Type Aliases

Blob

A Blob serialized as 0x-prefixed hex string

Bytes48

A commitment/proof serialized as 0x-prefixed hex string