alloy::consensus
Module transaction

Transaction types.

Modules

eip4844

[EIP-4844] constants, helpers, and types.

eip4844_utils

Utilities for working with EIP-4844 field elements and implementing SidecarCoder.

pooled

Defines the exact transaction variant that are allowed to be propagated over the eth p2p protocol.

signed_legacy_serde

Helper module for encoding signatures of transactions wrapped into Signed in legacy format.

Structs

BlobTransactionSidecar

This represents a set of blobs, and its corresponding commitments and proofs.

Recovered

Signed transaction with recovered signer.

SidecarBuilder

Build a BlobTransactionSidecar from an arbitrary amount of data.

SimpleCoder

Simple coder that only uses the last 31 bytes of each blob. This is the default coder for the SidecarBuilder.

TransactionInfo

Additional fields in the context of a (maybe) pending block that contains this transaction.

TransactionMeta

Additional fields in the context of a block that contains this mined transaction.

TxEip1559

A transaction with a priority fee (EIP-1559).

TxEip2930

Transaction with an AccessList (EIP-2930).

TxEip4844

EIP-4844 Blob Transaction

TxEip7702

A transaction with a priority fee (EIP-7702).

TxEip4844WithSidecar

EIP-4844 Blob Transaction

TxLegacy

Legacy transaction.

Enums

PooledTransaction

All possible transactions that can be included in a response to GetPooledTransactions. A response to GetPooledTransactions. This can include either a blob transaction, or a non-4844 signed transaction.

TxEip4844Variant

EIP-4844 Blob Transaction

TxEnvelope

The Ethereum EIP-2718 Transaction Envelope.

TxType

Ethereum TransactionType flags as specified in EIPs 2718, 1559, 2930, 4844, and 7702.

TypedTransaction

The TypedTransaction enum represents all Ethereum transaction request types.

Traits

SidecarCoder

A strategy for coding and decoding data into sidecars.

SignableTransaction

A signable transaction.

SignerRecoverable

A type that can recover the signer of a transaction.

Transaction

Represents a minimal EVM transaction.

Typed2718

A trait that helps to determine the type of the transaction.

Functions

from_eip155_value

Helper for decoding EIP-155 v value into y_parity boolean and optional chain_id.

to_eip155_value

Helper for encoding y_parity boolean and optional chain_id into EIP-155 v value.

Type Aliases

Blob

A Blob serialized as 0x-prefixed hex string

Bytes48

A commitment/proof serialized as 0x-prefixed hex string