alloy::contract

Trait CallDecoder

pub trait CallDecoder: Sealed { }

A trait for decoding the output of a contract function.

This trait is sealed and cannot be implemented manually. It is an implementation detail of CallBuilder.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.