alloy::contract
Trait CallDecoder

pub trait CallDecoder: Sealed { }

A trait for decoding the output of a contract function.

This trait is sealed and cannot be implemented manually. It is an implementation detail of CallBuilder.

Object Safety

This trait is not object safe.