alloy::eips

Module eip7623

Constants and utils for calldata cost.

See also EIP-7623: Increase calldata cost

Constants

NON_ZERO_BYTE_DATA_COST

The cost of a non-zero byte in calldata.

NON_ZERO_BYTE_MULTIPLIER

The multiplier for a non zero byte in calldata.

STANDARD_TOKEN_COST

The standard cost of calldata token.

TOTAL_COST_FLOOR_PER_TOKEN

The cost floor per token

Functions

tokens_in_calldata

Retrieve the total number of tokens in calldata.

transaction_floor_cost

Calculate the transaction cost floor as specified in EIP-7623.