alloy::network::eip2718

Trait Eip2718Envelope

pub trait Eip2718Envelope: Decodable2718 + Encodable2718 { }

An EIP-2718 envelope, blanket implemented for types that impl Encodable2718 and Decodable2718.

This envelope is a wrapper around a transaction, or a receipt, or any other type that is differentiated by an EIP-2718 transaction type.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.