alloy
Module providers

Interface with an Ethereum blockchain.

See alloy_provider for more details.

Modules

ext

Extended APIs for the provider module.

fillers

Transaction Fillers

layers

Useful layer implementations for the provider. Currently this module contains the AnvilLayer, AnvilProvider and ChainLayer types.

network

alloy-networks

utils

Provider-related utilities.

Structs

EthCall

A builder for an "eth_call" request. This type is returned by the Provider::call method.

EthCallParams

The parameters for an "eth_call" RPC request.

Identity

An identity layer that does nothing.

ParamsWithBlock

Helper struct that houses the params along with the BlockId.

PendingTransaction

Represents a transaction that is yet to be confirmed a specified number of times.

PendingTransactionBuilder

A builder for configuring a pending transaction watcher.

PendingTransactionConfig

Configuration for watching a pending transaction.

ProviderBuilder

A builder for constructing a Provider from various layers.

RootProvider

The root provider manages the RPC client and the heartbeat. It is at the base of every provider stack.

RpcWithBlock

A struct that takes an optional BlockId parameter.

Stack

A stack of two providers.

Enums

PendingTransactionError

Errors which may occur when watching a pending transaction.

ProviderCall

The primary future type for the Provider.

WatchTxError

Errors which may occur in heartbeat when watching a transaction.

Traits

Caller

Trait that helpes convert EthCall into a ProviderCall.

Network

Captures type info for network-specific RPC requests/responses.

Provider

Provider is parameterized with a network and a transport. The default transport is type-erased, but you can do Provider<Http, N>.

ProviderLayer

A layering abstraction in the vein of tower::Layer

WalletProvider

Trait for Providers, Fill stacks, etc, which contain NetworkWallet.

Functions

builder

Helper function to directly access ProviderBuilder with minimal generics.

Type Aliases

FilterPollerBuilder

A task that polls the provider with eth_getFilterChanges, returning a list of R.

ReqwestProvider

Type alias for a RootProvider using the Http transport and a reqwest client.