alloy

Module providers

Interface with an Ethereum blockchain.

See alloy_provider for more details.

Modules

bindings

Multicall bindings

ext

Extended APIs for the provider module.

fillers

Transaction Fillers

layers

Useful layer implementations for the provider.

mock

Mock transport and utility types.

network

alloy-networks

utils

Provider-related utilities.

Structs

CallItem

A singular call type that is mapped into aggregate, aggregate3, aggregate3Value call structs via the CallInfoTrait trait.

CallItemBuilder

Helper type to build a CallItem

DynProvider

A wrapper struct around a type erased Provider.

Dynamic

Marker for Dynamic Calls i.e where in SolCall type is locked to one specific type and multicall returns a Vec of the corresponding return type instead of a tuple.

Empty

Type indicating that the MulticallBuilder is empty.

EthCall

A builder for an "eth_call" request. This type is returned by the Provider::call method.

EthCallMany

A builder for an "eth_callMany" RPC request.

EthCallManyParams

The parameters for an "eth_callMany" RPC request.

EthCallParams

The parameters for an "eth_call" RPC request.

EthGetBlock

A builder for an "eth_getBlockByHash" request. This type is returned by the Provider::call method.

EthGetBlockParams

The parameters for an eth_getBlockBy{Hash, Number} RPC request.

Failure

A struct representing a failure in a multicall

Identity

An identity layer that does nothing.

MulticallBuilder

A Multicall3 builder

ParamsWithBlock

Helper struct that houses the params along with the BlockId.

PendingTransaction

Represents a transaction that is yet to be confirmed a specified number of times.

PendingTransactionBuilder

A builder for configuring a pending transaction watcher.

PendingTransactionConfig

Configuration for watching a pending transaction.

ProviderBuilder

A builder for constructing a Provider from various layers.

RootProvider

The root provider manages the RPC client and the heartbeat. It is at the base of every provider stack.

RpcWithBlock

A struct that takes an optional BlockId parameter.

SendableTxErr

Error when converting a SendableTx.

Stack

A stack of two providers.

WatchBlocks

A builder type for polling new blocks using the FilterPollerBuilder.

Web3Signer

A remote signer that leverages the underlying provider to sign transactions using "eth_signTransaction" requests.

Enums

MulticallError

Multicall errors.

PendingTransactionError

Errors which may occur when watching a pending transaction.

ProviderCall

The primary future type for the Provider.

SendableTx

A transaction that can be sent. This is either a builder or an envelope.

WatchTxError

Errors which may occur in heartbeat when watching a transaction.

Constants

MULTICALL3_ADDRESS

Default address for the Multicall3 contract on most chains. See: https://github.com/mds1/multicall

Traits

CallInfoTrait

A trait for converting CallItem into relevant call types.

Caller

Trait that helpes convert EthCall into a ProviderCall.

MulticallItem

A trait that is to be implemented by a type that can be distilled to a singular contract call item.

Network

Captures type info for network-specific RPC requests/responses.

Provider

Ethereum JSON-RPC interface.

ProviderLayer

A layering abstraction in the vein of tower::Layer

WalletProvider

Trait for Providers, Fill stacks, etc, which contain NetworkWallet.

Functions

builder

Helper function to directly access ProviderBuilder with minimal generics.

Type Aliases

BoxedFut

Boxed future type used in ProviderCall for non-wasm targets.

FilterPollerBuilder

A task that polls the provider with eth_getFilterChanges, returning a list of R.

ReqwestProviderDeprecated

Type alias for a RootProvider using the Http transport and a reqwest client.

Result

Result type for multicall operations.