Module alloy::pubsub

Available on crate feature pubsub only.

Ethereum JSON-RPC publish-subscribe tower service and type definitions.

You will likely not need to use this module; see the providers module for high-level usage of pubsub.

See alloy_pubsub for more details.

Structs

ConnectionHandle

A handle to a backend. Communicates to a ConnectionInterface on the backend.

ConnectionInterface

The reciprocal of ConnectionHandle.

PubSubFrontend

A PubSubFrontend is Transport composed of a channel to a running PubSub service.

RawSubscription

A Subscription is a feed of notifications from the server, identified by a local ID.

Subscription

A Subscription is a feed of notifications from the server of a specific type T, identified by a local ID.

Enums

SubscriptionItem

An item in a typed Subscription. This is either the expected type, or some serialized value of another type.

Traits

PubSubConnect

Configuration objects that contain connection details for a backend.