alloy::signers::k256::ecdsa

Module hazmat

Low-level ECDSA primitives.

⚠️ Warning: Hazmat!

YOU PROBABLY DON’T WANT TO USE THESE!

These primitives are easy-to-misuse low-level interfaces.

If you are an end user / non-expert in cryptography, do not use these! Failure to use them correctly can lead to catastrophic failures including FULL PRIVATE KEY RECOVERY!

Traits

DigestPrimitive

Bind a preferred Digest algorithm to an elliptic curve type.

SignPrimitive

Try to sign the given prehashed message using ECDSA.

VerifyPrimitive

Verify the given prehashed message using ECDSA.

Functions

bits2field

Partial implementation of the bits2int function as defined in RFC6979  2.3.2 as well as SEC1  2.3.8.

sign_prehashed

Sign a prehashed message digest using the provided secret scalar and ephemeral scalar, returning an ECDSA signature.

verify_prehashed

Verify the prehashed message against the provided ECDSA signature.