Trait alloy::signers::k256::ecdsa::signature::Verifier

pub trait Verifier<S> {
    // Required method
    fn verify(&self, msg: &[u8], signature: &S) -> Result<(), Error>;
}

Available on crate features signers and ecdsa-core only.

Verify the provided message bytestring using Self (e.g. a public key)

Required Methods

fn verify(&self, msg: &[u8], signature: &S) -> Result<(), Error>

Use Self to verify that the provided signature for a given message bytestring is authentic.

Returns Error if it is inauthentic, or otherwise returns ().