Trait alloy::signers::k256::elliptic_curve::pkcs8::der::referenced::RefToOwned

pub trait RefToOwned<'a> {
    type Owned: OwnedToRef<Borrowed<'a> = Self>
       where Self: 'a;

    // Required method
    fn ref_to_owned(&self) -> Self::Owned;
}

Available on crate feature signers only.

A trait for cloning a referenced structure and getting owned objects

This is the pendant to OwnedToRef

Required Associated Types

type Owned: OwnedToRef<Borrowed<'a> = Self> where Self: 'a

The resulting type after obtaining ownership.

Required Methods

fn ref_to_owned(&self) -> Self::Owned

Creates a new object taking ownership of the data