Trait alloy::signers::k256::elliptic_curve::pkcs8::der::Encode

pub trait Encode {
    // Required methods
    fn encoded_len(&self) -> Result<Length, Error>;
    fn encode(&self, encoder: &mut impl Writer) -> Result<(), Error>;

    // Provided methods
    fn encode_to_slice<'a>(&self, buf: &'a mut [u8]) -> Result<&'a [u8], Error> { ... }
    fn encode_to_vec(&self, buf: &mut Vec<u8>) -> Result<Length, Error> { ... }
    fn to_der(&self) -> Result<Vec<u8>, Error> { ... }
}

Available on crate feature signers only.

Encoding trait.

Required Methods

fn encoded_len(&self) -> Result<Length, Error>

Compute the length of this value in bytes when encoded as ASN.1 DER.

fn encode(&self, encoder: &mut impl Writer) -> Result<(), Error>

Encode this value as ASN.1 DER using the provided Writer.

Provided Methods

fn encode_to_slice<'a>(&self, buf: &'a mut [u8]) -> Result<&'a [u8], Error>

Encode this value to the provided byte slice, returning a sub-slice containing the encoded message.

fn encode_to_vec(&self, buf: &mut Vec<u8>) -> Result<Length, Error>

Available on crate feature alloc only.

Encode this message as ASN.1 DER, appending it to the provided byte vector.

fn to_der(&self) -> Result<Vec<u8>, Error>

Available on crate feature alloc only.

Encode this type as DER, returning a byte vector.

Object Safety

This trait is not object safe.