alloy::signers::utils

Function raw_public_key_to_address

pub fn raw_public_key_to_address(pubkey: &[u8]) -> Address

Convert a raw, uncompressed public key to its corresponding Ethereum address.

Warning

This method does not verify that the public key is valid. It is the caller’s responsibility to pass a valid public key. Passing an invalid public key will produce an unspendable output.

Panics

This function panics if the input is not exactly 64 bytes.