alloy::sol_types

Module sol_data

Solidity types.

These are the types that are built into Solidity.

See SolType for more details.

Structs

Address

Address - address

Array

Array - T[]

Bool

Bool - bool

ByteCount

Specifies the number of bytes in a FixedBytes array as a type.

Bytes

Bytes - bytes

FixedArray

FixedArray - T[M]

FixedBytes

FixedBytes - bytesX

Function

Function - function

Int

Int - intX

IntBitCount

Specifies the number of bits in an Int or Uint as a type.

String

String - string

Uint

Uint - uintX

Traits

SupportedFixedBytes

Statically guarantees that a FixedBytes byte count is marked as supported.

SupportedInt

Statically guarantees that a Int or Uint bit count is marked as supported.