alloy

Module transports

Low-level Ethereum JSON-RPC transport abstraction and implementations.

You will likely not need to use this module; see the providers module for high-level usage of transports.

See alloy_transport for more details.

Modules

http

alloy-transport-http

layers

Module for housing transport layers.

mock

Mock transport and utility types.

utils

Misc. utilities for building transports.

Macros

impl_future

Macro that return an impl Future type, with a Send bound on non-wasm targets.

Structs

BoxTransport

A boxed, Clone-able Transport trait object.

DualTransport

A transport that dispatches requests to one of two inner transports based on a handler.

HttpError

Type for holding HTTP errors such as 429 rate limit error.

Enums

Authorization

Basic, bearer or raw authentication in http or websocket transport.

RpcError

An RPC error.

TransportErrorKind

Transport error.

Traits

DualTransportHandler

Trait that determines how to dispatch a request given two transports.

IntoBoxTransport

Trait for converting a transport into a boxed transport.

Transport

A Transport manages the JSON-RPC request/response lifecycle.

TransportConnect

Connection details for a transport.

Type Aliases

BoxFuture

Type alias for a pin-boxed future, with a Send bound on non-wasm targets.

Pbf

Pin-boxed future.

RpcFut

Future for RPC-level requests.

RpcResult

The result of a JSON-RPC request.

TransportError

A transport error is an RpcError containing a TransportErrorKind.

TransportFut

Future for transport-level requests.

TransportResult

A transport result is a Result containing a TransportError.