alloy::transports

Module mock

Mock transport and utility types.

MockTransport returns responses that have been pushed into its associated Asserter’s queue using FIFO.

Examples

use alloy_transport::mock::*;

let asserter = Asserter::new();
let provider = ProviderBuilder::new()
    /* ... */
    .on_mocked_client(asserter.clone());

let n = 12345;
asserter.push_success(&n);
let actual = provider.get_block_number().await.unwrap();
assert_eq!(actual, n);

Structs

Asserter

Container for pushing responses into a MockTransport.

MockTransport

A transport that returns responses from an associated Asserter.

Type Aliases

MockResponse

A mock response that can be pushed into an Asserter.