Trait alloy::transports::BoxTransportConnect

pub trait BoxTransportConnect {
    // Required methods
    fn is_local(&self) -> bool;
    fn get_boxed_transport<'a, 'b>(
        &'a self
    ) -> Pin<Box<dyn Future<Output = Result<BoxTransport, RpcError<TransportErrorKind>>> + Send + 'b>>
       where 'a: 'b;
}

Available on crate feature transports only.

Connection details for a transport that can be boxed.

This trait is implemented for TransportConnect implementors that produce a boxable transport. It can be used to create a boxed transport without knowing the exact type of the transport.

This trait is separate from TransportConnect`` to hide the associated type in when this trait is a trai object. It is intended to allow creation of several unlike transports or clients at once. E.g. in something like Vec<&dyn BoxTransportConnect>.

Required Methods

fn is_local(&self) -> bool

Returns true if the transport is a local transport.

fn get_boxed_transport<'a, 'b>( &'a self ) -> Pin<Box<dyn Future<Output = Result<BoxTransport, RpcError<TransportErrorKind>>> + Send + 'b>>
where 'a: 'b,

Connect to a transport, and box it.